% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{agrupar_casos_eventos}
\alias{agrupar_casos_eventos}
\title{Agrupar por tipo de enfermedad o evento}
\usage{
agrupar_casos_eventos(data_event, nomb_col = "cod_eve")
}
\arguments{
\item{data_event}{Un `data.frame` que contiene los datos de la
enfermedad o evento}

\item{nomb_col}{Un `character` (cadena de caracteres) con el nombre de
la columna en los datos de la enfermedad o evento que contiene los códigos
de los eventos; su valor por defecto es `"cod_eve"`}
}
\value{
Un `data.frame` con los datos de la enfermedad o evento agrupados
por sus tipos
}
\description{
Función que agrupa los casos por tipo de enfermedad o evento
}
\examples{
data(dengue2020)
data_limpia <- limpiar_data_sivigila(data_event = dengue2020)
agrupar_casos_eventos(data_event = data_limpia,
                      nomb_col = "cod_eve")
}
