% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{agrupar_casos_semanaepi}
\alias{agrupar_casos_semanaepi}
\title{Agrupar por semana y casos}
\usage{
agrupar_casos_semanaepi(data_event)
}
\arguments{
\item{data_event}{Un data frame que contiene los datos de
una enfermedad o evento}
}
\value{
Un data frame con los datos de una enfermedad o
evento agrupados por semana epidemiológica y número de casos
}
\description{
Función que agrupa los datos de una enfermedad o evento
por semana epidemiológica y número de casos
}
\examples{
data_event <- import_data_event(2019, "DENGUE")
data_event <- limpiar_data_sivigila(data_event)
agrupar_casos_semanaepi(data_event = data_event)
}
