% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{agrupar_dpto}
\alias{agrupar_dpto}
\title{Agrupar por departamento y casos}
\usage{
agrupar_dpto(data_event, col_nombre = "cod_dpto_o", porcentaje = FALSE)
}
\arguments{
\item{data_event}{Los datos de la enfermedad}

\item{col_nombre}{Nombre de la columna en los datos de la
enfermedad que contiene los códigos de departamento}

\item{porcentaje}{Indica si se requiere agregar un porcentaje
de casos como columna}
}
\value{
Los datos de la enfermedad agrupados por códigos de
departamento y número de casos
}
\description{
Función que agrupa los datos por códigos de departamento y
número de casos
}
\examples{
data_event <- import_data_event(2019, "DENGUE")
data_event <- limpiar_encabezado(data_event)
agrupar_dpto(data_event, col_nombre = "cod_dpto_o", porcentaje = FALSE)
}
