% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{agrupar_edad}
\alias{agrupar_edad}
\title{Agrupar por edad y casos}
\usage{
agrupar_edad(
  data_event,
  nomb_col = "edad",
  interval_edad = 10,
  porcentaje = FALSE
)
}
\arguments{
\item{data_event}{Un `data.frame` que contiene los datos de la enfermedad
o evento}

\item{nomb_col}{Un `character` (cadena de caracteres) con el nombre
de la columna de los datos de la enfermedad o evento que contiene las edades;
su valor por defecto es `"edad"`}

\item{interval_edad}{Un `numeric` (numerico) que contiene el intervalo del
rango de edades; su valor por defecto es `10`}

\item{porcentaje}{Un `boolean` (TRUE o FALSE) que indica si
se debe agregar el porcentaje de casos como columna; su valor
por defecto es `FALSE`}
}
\value{
Un `data.frame` con los datos de la enfermedad o evento agrupados
por edad y número de casos
}
\description{
Función que agrupa los datos de una enfermedad o evento por edad
y número de casos
}
\examples{
data(dengue2020)
data_limpia <- limpiar_data_sivigila(data_event = dengue2020)
agrupar_edad(data_event = data_limpia,
             nomb_col = "edad",
             porcentaje = FALSE)
}
