% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{agrupar_fecha_inisintomas}
\alias{agrupar_fecha_inisintomas}
\title{Group by symptom onset date and cases}
\usage{
agrupar_fecha_inisintomas(data_event, col_nombre = "ini_sin", tipo = "month")
}
\arguments{
\item{data_event}{The disease data}

\item{col_nombre}{Column name in the disease data that contains
the symptom onset dates}

\item{tipo}{Time unit (day, month and year)}
}
\value{
The disease data grouped by symptom onset date and number of cases
}
\description{
Function that groups the disease data by symptom onset date and
number of cases
}
\examples{
data_event <- import_data_event(2019, "DENGUE")
data_event <- limpiar_encabezado(data_event)
agrupar_fecha_inisintomas(data_event, col_nombre = "ini_sin", tipo = "month")
}
