% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{agrupar_fecha_notifica}
\alias{agrupar_fecha_notifica}
\title{Agrupar por fecha de notificación y casos}
\usage{
agrupar_fecha_notifica(data_event, nomb_col = "fec_not")
}
\arguments{
\item{data_event}{Un `data.frame` que contiene los datos de la enfermedad
o evento}

\item{nomb_col}{Un `character` (cadena de caracteres) con el nombre de
la columna de los datos de la enfermedad o evento que contiene las
fechas de notificación; su valor por defecto es `"fec_not"`}
}
\value{
Un `data.frame` con los datos de enfermedades agrupados por fecha de
notificación y número de casos
}
\description{
Función que agrupa los datos de una enfermedad o evento por fecha de
notificación y número de casos
}
\examples{
data(dengue2020)
data_limpia <- limpiar_data_sivigila(data_event = dengue2020)
agrupar_fecha_notifica(data_event = data_limpia,
                       nomb_col = "fec_not")
}
