% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{agrupar_mun}
\alias{agrupar_mun}
\title{Agrupar por municipios y casos}
\usage{
agrupar_mun(
  data_event,
  dept_nombre = NULL,
  col_nombre = "cod_mun_o",
  porcentaje = FALSE
)
}
\arguments{
\item{data_event}{Los datos de la enfermedad}

\item{dept_nombre}{El nombre del departamento}

\item{col_nombre}{Nombre de la columna en los datos de la
enfermedad que contiene los códigos de municipios}

\item{porcentaje}{Indica si se requiere agregar un porcentaje
de casos como columna}
}
\value{
Los datos de la enfermedad agrupados por códigos de
municipios y número de casos
}
\description{
Función que agrupa los datos por códigos de municipios y número de casos
}
\examples{
data_event <- import_data_event(2019, "DENGUE")
data_event <- limpiar_encabezado(data_event)
agrupar_mun(data_event,
            dept_nombre = "Antioquia",
            col_nombre = "cod_mun_o",
            porcentaje = FALSE)
}
