% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{agrupar_pob_especial}
\alias{agrupar_pob_especial}
\title{Agrupar por población especial y casos}
\usage{
agrupar_pob_especial(data_event, col_pob = "poblacion", porcentaje = TRUE)
}
\arguments{
\item{data_event}{Un `data.frame` que contiene los datos de la enfermedad
o evento}

\item{col_pob}{Un `character` (cadena de caracteres) con el nombre de la
columna que contiene las poblaciones especiales en los datos de la
enfermedad o evento; su valor por defecto es `"poblacion"`}

\item{porcentaje}{Un `boolean` (TRUE o FALSE) que indica si se debe
agregar una columna con el porcentaje de casos; su valor por
defecto es `FALSE`}
}
\value{
Un `data.frame` con los datos de la enfermedad o evento agrupados
por poblaciones especiales y casos
}
\description{
Función que agrupa los datos de la enfermedad o evento por población
especial y casos
}
\examples{
data(dengue2020)
data_limpia <- limpiar_data_sivigila(data_event = dengue2020)
agrupar_pob_especial(data_event = data_limpia,
                     col_pob = "poblacion",
                     porcentaje = TRUE)
}
