% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{agrupar_rango_edad_casos}
\alias{agrupar_rango_edad_casos}
\title{Agrupar por rango de edad y casos}
\usage{
agrupar_rango_edad_casos(
  data_event,
  nomb_col,
  col_adicional = NULL,
  min_val,
  max_val,
  paso
)
}
\arguments{
\item{data_event}{Un `data.frame` que contiene los datos de la
enfermedad o evento}

\item{nomb_col}{Un `character` (cadena de caracteres) con
el nombre de la columna de los datos de la enfermedad o evento
que contiene las edades}

\item{col_adicional}{Un `character` (cadena de caracteres) con
el nombre adicional de la columna de los datos de la enfermedad
o evento para agrupar con la edad; su valor por defecto es
`NULL`}

\item{min_val}{Un `numeric` (numerico) que contiene el valor mínimo
de las edades}

\item{max_val}{Un `numeric` (numerico) que contiene el valor máximo
de las edades}

\item{paso}{Un `numeric` (numerico) que contiene el valor del paso
para generar el rango de edades}
}
\value{
Un `data.frame` con los datos de la enfermedad o evento
agrupados por el rango de edad y número de casos
}
\description{
Función que agrupa los datos de una enfermedad o evento por rango
de edad y número de casos
}
\examples{
data(dengue2020)
data_limpia <- limpiar_data_sivigila(data_event = dengue2020)
data_edad <- agrupar_cols_casos(data_event = data_limpia,
                                c("edad", "semana"),
                                porcentaje = TRUE)
agrupar_rango_edad_casos(data_event = data_edad,
                         nomb_col = "edad",
                         min_val = 0,
                         max_val = max(data_edad$edad),
                         paso = 10)
}
