% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{agrupar_sex}
\alias{agrupar_sex}
\title{Agrupar por sexo y casos}
\usage{
agrupar_sex(data_event, col_nombre = "sexo", porcentaje = TRUE)
}
\arguments{
\item{data_event}{Los datos de enfermedades}

\item{col_nombre}{Nombre de la columna en los datos de enfermedades
que contiene el sexo}

\item{porcentaje}{Indica si es necesario agregar un porcentaje de casos
como una columna}
}
\value{
Los datos de enfermedades agrupados por sexo y número de casos
}
\description{
Función que agrupa los datos de enfermedades
por sexo y número de casos
}
\examples{
data_event <- import_data_event(2019, "DENGUE")
data_event <- limpiar_encabezado(data_event)
agrupar_sex(data_event, col_nombre = "sexo", porcentaje = TRUE)
}
