% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{agrupar_sex_semanaepi}
\alias{agrupar_sex_semanaepi}
\title{Agrupar por sexo, semana epidemiológica y casos}
\usage{
agrupar_sex_semanaepi(
  data_event,
  col_nombres = c("sexo", "semana"),
  porcentaje = TRUE
)
}
\arguments{
\item{data_event}{Los datos de enfermedades}

\item{col_nombres}{Nombres de las columnas en los datos de enfermedades
que contienen el sexo y las semanas epidemiológicas}

\item{porcentaje}{Indica si es necesario agregar un porcentaje
de casos como una columna}
}
\value{
Los datos de enfermedades agrupados por
sexo, semana epidemiológica y número de casos
}
\description{
Función que agrupa los datos de enfermedades por sexo, semana
epidemiológica y número de casos
}
\examples{
data_event <- import_data_event(2019, "DENGUE")
data_event <- limpiar_encabezado(data_event)
agrupar_sex_semanaepi(data_event,
                      col_nombres = c("sexo", "semana"),
                      porcentaje = TRUE)
}
