% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{agrupar_tipo_caso}
\alias{agrupar_tipo_caso}
\title{Agrupar por la clasificacion inicial del caso}
\usage{
agrupar_tipo_caso(data_event, nomb_cols = "tip_cas")
}
\arguments{
\item{data_event}{Un `data.frame` que contiene los datos de la
enfermedad o evento}

\item{nomb_cols}{Un `character` (cadena de caracteres) con el nombre de
las columna(s) en los datos de la enfermedad o evento que contiene la
clasificacion inicial del caso; su valor por defecto es `"tip_cas"`}
}
\value{
Un `data.frame` con los datos de la enfermedad o evento agrupados
por la clasificacion inicial del caso y/u otras variables como años
}
\description{
Función que agrupa los casos por la clasificacion inicial del caso
}
\examples{
data(dengue2020)
data_limpia <- limpiar_data_sivigila(data_event = dengue2020)
agrupar_tipo_caso(data_event = data_limpia,
                  nomb_cols = "tip_cas")
}
