% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{calcular_incidencia_sex}
\alias{calcular_incidencia_sex}
\title{Calcular incidencia por sexo}
\usage{
calcular_incidencia_sex(
  data_incidencia,
  data_agrupada,
  year,
  dpto = NULL,
  mpio = NULL
)
}
\arguments{
\item{data_incidencia}{Un `data.frame` que contiene la proyecciones
poblaciones del DANE}

\item{data_agrupada}{Un `data.frame` que contiene los datos de la enfermedad
agrupados por departamento o municipio y número de casos}

\item{year}{Un `numeric` (numerico) con el año que se debe tomar de las
proyecciones poblacionales}

\item{dpto}{Un `character` (cadena de caracteres) o `numeric` (numérico)
que contiene el código o nombre del departamento; su valor por
defecto es `NULL`}

\item{mpio}{Un `character` (cadena de caracteres) o `numeric` (numérico)
que contiene el código o nombre del municipio; su valor por defecto es `NULL`}
}
\value{
Un `data.frame` con el calculo de la incidencia para todos los
departamentos de Colombia o los municipios de un departamento
}
\description{
Función que calcula la incidencia de una enfermedad o evento para todos los
departamentos de Colombia o los municipios de un departamento por sexo
}
\examples{
\dontrun{
data(dengue2020)
data_limpia <- limpiar_data_sivigila(data_event = dengue2020)
proyecciones <- import_data_incidencia()
data_agrupada <- agrupar_sex(data_limpia)
calcular_incidencia_sex(data_incidencia = proyecciones,
                        data_agrupada = data_agrupada,
                        dpto = "05",
                        year = 2020)
calcular_incidencia_sex(data_incidencia = proyecciones,
                        data_agrupada = data_agrupada,
                        dpto = "05",
                        mpio = "Medellin",
                        year = 2020)
}
}
