% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{clean_depto_codes}
\alias{clean_depto_codes}
\title{clean_depto_codes}
\usage{
clean_depto_codes(disease_data, col_data_codes, geo_data, col_geo_codes)
}
\arguments{
\item{disease_data}{The disease data}

\item{col_data_codes}{Column name of departments codes of the disease data}

\item{col_geo_codes}{Column name of departments codes of the geographical data}

\item{geo_codes}{Geographical data that includes the departments codes}
}
\value{
Clean codes of disease data departments
}
\description{
Función que limpia los codigos de los departamentos de los datos de la enfermedad
Function that cleans the departments codes of the disease data
}
\examples{
geo_codes <- import_geo_codes()
disease_data <-  import_data_disease_by_year(2019, "DENGUE")
clean_depto_codes(disease_data, col_data_codes = "cod_dpto_o", geo_codes, col_geo_codes = "c_digo_departamento")
}
