% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{clean_depto_disease_codes}
\alias{clean_depto_disease_codes}
\title{Clean Depto Disease Codes}
\usage{
clean_depto_disease_codes(deptos_data, disease_data, make_group = TRUE)
}
\arguments{
\item{deptos_data}{Departments data}

\item{disease_data}{Disease data}
}
\value{
The filtered data with the disease selected information
}
\description{
Función que limpia la información de la enfermedad respecto a los codigos de los departamentos
Function that cleans the disease data with respect to the codes of the departments
}
\examples{
clean_depto_disease_codes(deptos_data, disease_data)
}
