% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{clean_disease_ages}
\alias{clean_disease_ages}
\title{Clean ages of the disease data}
\usage{
clean_disease_ages(disease_data, col_name = "edad")
}
\arguments{
\item{disease_data}{The disease data}

\item{col_name}{Column name in the disease data that contains the ages}
}
\value{
The disease data with clean ages
}
\description{
Function that cleans ages of the disease data
}
\examples{
disease_data <- import_linelist_disease_year(2020, "DENGUE")
clean_disease_ages(disease_data, col_name = "edad")
}
