% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{clean_disease_cols}
\alias{clean_disease_cols}
\title{Clean outliers from disease data columns}
\usage{
clean_disease_cols(disease_data)
}
\arguments{
\item{disease_data}{The disease data}
}
\value{
The disease data with clean columns
}
\description{
Function that clean outliers from disease data columns
}
\examples{
disease_data <- import_linelist_disease_year(2020, "DENGUE")
disease_data <- clean_header(disease_data)
clean_disease_cols(disease_data)
}
