% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{clean_header}
\alias{clean_header}
\title{Clean the header labels of the disease data}
\usage{
clean_header(disease_data)
}
\arguments{
\item{disease_data}{The disease data}
}
\value{
Formatted header labels
}
\description{
Function that cleans the header labels of the disease data
}
\examples{
disease_data <- import_linelist_disease_year(2019, "DENGUE")
clean_header(disease_data)
}
