% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{clean_labels}
\alias{clean_labels}
\title{Estandarizar etiquetas}
\usage{
clean_labels(
  x,
  sep = "_",
  transformation = "Any-Latin; Latin-ASCII",
  protect = ""
)
}
\arguments{
\item{x}{Un vector de etiquetas, normalmente proporcionado como caracteres.}

\item{sep}{Una cadena de caracteres utilizada como separador, con un valor
predeterminado de '_'.}

\item{transformation}{Una cadena que se pasa a
\code{stringi::stri_trans_general()} para la conversión. Por defecto es
"Any-Latin; Latin-ASCII", lo que convierte cualquier carácter no latino a
caracteres latinos y luego convierte todos los caracteres acentuados a
caracteres ASCII. Consulta \code{stringi::stri_trans_list()} para una lista
completa de opciones.}

\item{protect}{Una cadena de caracteres que define la puntuación que se debe
proteger. Esto ayuda a evitar que se eliminen símbolos significativos
como > y <.}
}
\description{
Código tomado de la función \code{epitrix::clean_labels()}.
La función \code{clean_labels()} de \pkg{epitrix} se reutiliza con
permiso y está lincenciada bajo MIT al igual que \pkg{epitrix}.
\pkg{epitrix} está en CRAN.
}
\note{
Debido a diferencias entre el motor de transliteración subyacente
(ICU), las transformaciones predeterminadas no transliterarán correctamente
las diéresis alemanas (umlaute). Puedes agregarlas especificando "de-ASCII"
en la cadena \code{transformation} después de "Any-Latin".

El código original fue escrito por los autores de \pkg{epitrix}.
Consulta \url{https://CRAN.R-project.org/package=epitrix} para más detalles.
}
\examples{
\dontrun{
clean_labels("-_-Esto es; Una    Fräse**./extraña...")
clean_labels("-_-Esto es; Una    Fräse**./extraña...", sep = ".")
input <- c("Pedro y stëven",
           "pedro-y.stëven",
           "pëtêr y stëven  _-")
clean_labels(input)
# No transliterar palabras no latinas
clean_labels(input, transformation = "Latin-ASCII")
# proteger símbolos útiles
clean_labels(c("energía > 9000", "energía < 9000"), protect = "><")
# si solo deseas limpiar acentos, transformar a minúsculas y transliterar,
# puedes especificar "[:punct:][:space:]" para protect:
clean_labels(input, protect = "[:punct:][:space:]")
}
}
\keyword{internal}
