% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{concatenar_vals_token}
\alias{concatenar_vals_token}
\title{Concatenar vals con separador o token}
\usage{
concatenar_vals_token(
  vals,
  longitud = 3,
  princ_token = ", ",
  final_token = "y "
)
}
\arguments{
\item{vals}{Los vals}

\item{longitud}{Longitud de los vals}

\item{princ_token}{Separador o token principal}

\item{final_token}{Separador o token final}
}
\value{
Valor final concatenado
}
\description{
Función que concatena vals con un separador o token específico
}
\examples{
concatenar_vals_token(vals = c("enero", "febrero", "marzo"),
                         longitud = 3,
                         princ_token = ", ",
                         final_token = "y ")
}
