% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{concatenate_values_token}
\alias{concatenate_values_token}
\title{Concatenate values with separator or token}
\usage{
concatenate_values_token(
  values,
  length = 3,
  main_token = ", ",
  final_token = "y "
)
}
\arguments{
\item{values}{The values}

\item{length}{Values length}

\item{main_token}{Main separator or token}

\item{final_token}{Final separator or token}
}
\value{
Concatenated final value
}
\description{
Function that concatenates values with a specific separator or token
}
\examples{
concatenate_values_token(values = c("enero", "febrero", "marzo"),
                         length = 3,
                         main_token = ", ",
                         final_token = "y ")
}
