% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{concatenate_values_with_token}
\alias{concatenate_values_with_token}
\title{concatenate_values_with_token}
\usage{
concatenate_values_with_token(
  values,
  length = 3,
  main_token = ", ",
  final_token = "y "
)
}
\arguments{
\item{values}{Values}

\item{length}{Length}

\item{main_token}{Main token}

\item{final_token}{Final token}
}
\value{
Concatenated final value
}
\description{
Función que concantena valores con un separador o token especifico
Function that concatenates values with a specific separator or token
}
\examples{
concatenate_values_with_token(values = c("enero", "febrero", "marzo"), length = 3, main_token = ", ", final_token = "y ") 
}
