% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dengue2020.R
\docType{data}
\name{dengue2020}
\alias{dengue2020}
\title{Datos Dengue 2020 del SIVIGILA en sivirep}
\format{

}
\usage{
data(dengue2020)
}
\description{
Datos obtenidos del SIVIGILA
(Sistema de Vigilancia en Salud Pública de Colombia)
}
\examples{
data(dengue2020)
}
\keyword{datasets}
