% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divipoladata.R
\docType{data}
\name{divipoladata}
\alias{divipoladata}
\title{Códigos e información geografica del DIVIPOLA en sivirep}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
divipoladata
}
\description{
Data from API de Datos Abiertos de Colombia
}
\examples{
divipoladata
}
\keyword{datasets}
