% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estandarizar_texto}
\alias{estandarizar_texto}
\title{Estandarizar un texto con tíldes}
\usage{
estandarizar_texto(text)
}
\arguments{
\item{text}{Un `character` (cadena de caracteres) que contiene
un texto con tíldes}
}
\value{
Un `character` (cadena de caracteres) con el texto sin
tíldes
}
\description{
Función que estandariza un texto que contiene tíldes
}
\examples{
estandarizar_texto(text = "MEDELLÍN")
}
