% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{filtrar_event}
\alias{filtrar_event}
\title{Filtrar enfermedad}
\usage{
filtrar_event(nombre_event, data_sivigila)
}
\arguments{
\item{nombre_event}{Nombre de la enfermedad}

\item{data_sivigila}{El conjunto de datos}
}
\value{
Datos filtrados con la enfermedad seleccionada
}
\description{
Función que filtra por nombre de enfermedad en un conjunto de datos
}
\examples{
data_sivigila <- import_data_resumen_sivigila()
data_sivigila <- limpiar_encabezado(data_sivigila)
filtrar_event("MALAR", data_sivigila)
}
