% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{find_name_file_path}
\alias{find_name_file_path}
\title{Find Name File Path or URL}
\usage{
find_name_file_path(path)
}
\arguments{
\item{path}{The path or URL}
}
\value{
The name file
}
\description{
Función que obtiene el nombre del archivo a descargar desde una URL o ruta
Function that obtains the name of the file to download from a URL or path
}
\examples{
find_name_file_path("DENGUE")
}
