% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{format_cod_geo}
\alias{format_cod_geo}
\title{Formatear código geográfico}
\usage{
format_cod_geo(cod_geo, etiqueta, digitos, tam)
}
\arguments{
\item{cod_geo}{Un `numeric` (númerico) o `character`
(cadena de caracteres) que contiene el código geográfico.}

\item{etiqueta}{Un `character` (cadena de caracteres) con el nombre
de la etiqueta de la validación relacionada a la longitud máxima del
código geográfico; se refiere al tipo de división geográfica ("municipio",
"departamento").}

\item{digitos}{Un `numeric` (númerico) que contiene el número de digitos
que debe tener individualmente el código geográfico.}

\item{tam}{Un `numeric` (númerico) que contiende el tamaño o la longitud
máxima que debe tener el código geográfico.}
}
\value{
Un `character` (cadena de caracteres) con el código geográfico
formateado.
}
\description{
Función que da el formato deseado a un código geográfico.
}
\keyword{internal}
