% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{format_dates_values}
\alias{format_dates_values}
\title{Format dates}
\usage{
format_dates_values(
  disease_data,
  date_format = "\%AAAA-\%MM-\%DD",
  col_names = c()
)
}
\arguments{
\item{disease_data}{The disease data}

\item{date_format}{Date format}

\item{col_names}{Column names to format}
}
\value{
The data with formatted dates
}
\description{
Function that gives a specific format to a date
}
\examples{
disease_data <- import_linelist_disease_year(2020, "DENGUE")
remove_error_dates(disease_data, 
                   date_format = "\%AAAA-\%MM-\%DD", 
                   col_names = c("ini_sin", "fec_hos"))
}
