% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{geo_filter}
\alias{geo_filter}
\title{Filter by departments and municipalities}
\usage{
geo_filter(disease_data, dept_name = NULL, mun_name = NULL)
}
\arguments{
\item{dept_name}{The department name}

\item{mun_name}{The municipality name}

\item{name_disease}{The disease name}
}
\value{
Data filtered with the disease, departments and
municipalities selected
}
\description{
Function that filters disease data by departments and municipalities
}
\examples{
disease_data <- import_linelist_disease_year(2019, "DENGUE")
disease_data <- clean_header(disease_data)
geo_filter(disease_data, dept_name = "ANTIOQUIA")
}
