% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{geographical_filter}
\alias{geographical_filter}
\title{Filter by departments and municipalities}
\usage{
geographical_filter(disease_data, department = NULL, municipalitie = NULL)
}
\arguments{
\item{name_disease}{Name of the disease}

\item{departments}{Names of departments}

\item{municipalities}{Names of departments}
}
\value{
Data filtered with the disease selected
}
\description{
Function that filters disease data by departments and municipalities
}
\examples{
disease_data <- import_linelist_disease_year(2019, "DENGUE")
disease_data <- clean_header(disease_data)
geographical_filter(disease_data, department = "ANTIOQUIA")
}
