% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_section.R
\name{get_cases_distribution_by_age_and_gender_section}
\alias{get_cases_distribution_by_age_and_gender_section}
\title{get_cases_distribution_by_age_and_gender_section}
\usage{
get_cases_distribution_by_age_and_gender_section(
  disease_data,
  year,
  col_names = c("EDAD", "SEXO"),
  percentage = T
)
}
\arguments{
\item{disease_data}{Disease data}

\item{year}{Year}

\item{col_names}{Data set column names}

\item{percentage}{Percentage}
}
\value{
A list with the cases by age range, the age range major cases, the percentage major cases and the section plot
}
\description{
Función que genera la seccion de distribucion de casos por rango de edad y genero
Function that generates the section of cases distribution by age range and gender
}
\examples{
disease_data <-  import_data_disease_by_year(2020, "DENGUE")
distribution_by_age_and_gender_section <- get_cases_distribution_by_age_and_gender_section(disease_data, year = 2020, col_names = c("EDAD", "SEXO"), percentage = T)
}
