% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_section.R
\name{get_cases_distribution_by_notification_date_section}
\alias{get_cases_distribution_by_notification_date_section}
\title{get_cases_distribution_by_notification_date_section}
\usage{
get_cases_distribution_by_notification_date_section(
  disease_data,
  year,
  type = "month",
  col_name = "FEC_NOT"
)
}
\arguments{
\item{disease_data}{Disease data}

\item{year}{Year}

\item{type}{Time unit}

\item{col_name}{Data set column name}
}
\value{
A list with the cases by notification date and the section plot
}
\description{
Función que genera la seccion de distribucion de casos por fecha de notificacion
Function that generates the section of cases distribution by notification date
}
\examples{
disease_data <-  import_data_disease_by_year(2020, "DENGUE")
distribution_by_notification_date_section <- get_cases_distribution_by_notification_date_section(disease_data, year = 2020, type = "month", col_name = "FEC_NOT")
}
