% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_section.R
\name{get_cases_distribution_by_onset_sym_section}
\alias{get_cases_distribution_by_onset_sym_section}
\title{get_cases_distribution_by_onset_sym_section}
\usage{
get_cases_distribution_by_onset_sym_section(
  disease_data,
  year,
  type = "month",
  col_name = "INI_SIN",
  col_cmp = "FEC_HOS"
)
}
\arguments{
\item{disease_data}{Disease data}

\item{year}{Year}

\item{type}{Time unit}

\item{col_name}{Data set column name}

\item{col_cmp}{Data set column compare}
}
\value{
A list with the cases by onset symptoms date and the section plot
}
\description{
Función que genera la seccion de distribucion de casos por fecha de inicio de sintomas
Function that generates the section of distribution of cases by onset symptoms date
}
\examples{
disease_data <-  import_data_disease_by_year(2020, "DENGUE")
get_cases_distribution_by_onset_sym_section(disease_data, year = 2020, type = "month", col_name = "INI_SIN", col_cmp = "FEC_HOS")
}
