% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_section.R
\name{get_cases_distribution_by_special_population_section}
\alias{get_cases_distribution_by_special_population_section}
\title{get_cases_distribution_by_special_population_section}
\usage{
get_cases_distribution_by_special_population_section(
  disease_data,
  year,
  col_name = "Poblacion",
  percentage = F,
  plot_title
)
}
\arguments{
\item{disease_data}{Disease data}

\item{year}{Year}

\item{col_name}{Data set column name}

\item{percentage}{Percentage}
}
\value{
A list with the cases by special population and the section plot
}
\description{
Función que genera la seccion de distribucion de casos por poblacion especial
Function that generates the section of cases distribution by special population
}
\examples{
disease_data <-  import_data_disease_by_year(2020, "DENGUE")
distribution_by_special_population_section <- get_cases_distribution_by_special_population_section(disease_data, year = 2020, col_name = "Poblacion", percentage = F)
}
