% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_section.R
\name{get_cases_distribution_spatial_section}
\alias{get_cases_distribution_spatial_section}
\title{get_cases_distribution_spatial_section}
\usage{
get_cases_distribution_spatial_section(
  disease_data,
  year,
  col_name = "COD_DPTO_R",
  percentage = F,
  plot_title = ""
)
}
\arguments{
\item{disease_data}{Disease data}

\item{year}{Year}

\item{col_name}{Data set column name}

\item{percentage}{Percentage}

\item{plot_title}{Plot title}
}
\value{
A list with the cases by deptos and the section map
}
\description{
Función que genera la seccion de distribucion espacial de casos
Function that generates the section of cases distribution spatial
}
\examples{
disease_data <-  import_data_disease_by_year(2020, "DENGUE")
distribution_by_special_population_section <- get_cases_distribution_spatial_section(disease_data, year = 2020, col_name = "Poblacion", percentage = F)
}
