% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_departments}
\alias{get_departments}
\title{Get departments of Colombia}
\usage{
get_departments()
}
\value{
The departments of Colombia
}
\description{
Function that gets the departments of Colombia
}
\examples{
get_departments()
}
