% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_depto_names}
\alias{get_depto_names}
\title{get_depto_names}
\usage{
get_depto_names(disease_data)
}
\arguments{
\item{disease_data}{Disease data}
}
\value{
dataset with the department names
}
\description{
Función que obtiene los nombres de los departamentos
Function that gets the department names
}
\examples{
disease_data <- import_data_disease_by_year(2020, "DENGUE")
get_depto_names(disease_data)
}
