% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_geo_occurrence_type}
\alias{get_geo_occurrence_type}
\title{Get geographic information of the disease data}
\usage{
get_geo_occurrence_type(code_disease)
}
\arguments{
\item{code_disease}{The disease code}
}
\value{
The geographic codes of the disease data
}
\description{
Function that standardizes the geographic codes of the disease data
}
\examples{
get_geo_occurrence_type(code_disease = 210)
}
