% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_info_depts}
\alias{get_info_depts}
\title{Get geographic information of the disease data}
\usage{
get_info_depts(department = NULL, municipality = NULL)
}
\arguments{
\item{department}{The department name}

\item{municipality}{The municipality data}
}
\value{
The geographic information of the disease data
}
\description{
Function that gets the geographic information of the disease data
}
\examples{
get_info_depts(department = "ANTIOQUIA")
}
