% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_major_cases}
\alias{get_major_cases}
\title{get_row_major_cases}
\usage{
get_major_cases(disease_data, col_name = "casos", percentage = T)
}
\arguments{
\item{disease_data}{Disease data}

\item{col_name}{Data set column name}

\item{percentage}{Percentage value}
}
\value{
row with major cases
}
\description{
Función que obtiene la fila con mayor numero de casos
Function that gets the row with the major number of cases
}
\examples{
disease_data <- import_data_disease_by_year(2020, "DENGUE")
get_major_cases(disease_data, col_name = "casos", percentage = T)
}
