% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_months_major_cases}
\alias{get_months_major_cases}
\title{get_months_major_cases}
\usage{
get_months_major_cases(
  disease_data,
  col_dates,
  col_cases,
  top = 3,
  concat_values = T
)
}
\arguments{
\item{disease_data}{Disease data}

\item{col_dates}{Data set column dates name}

\item{col_cases}{Data set column cases name}

\item{top}{Top}

\item{concat_values}{Concatenate values}
}
\value{
months major cases
}
\description{
Función que obtiene los meses con la cantidad mayor de casos
Function that gets the months with the largest number of cases
}
\examples{
disease_data <- import_data_disease_by_year(2020, "DENGUE")
clean_disease_ages(disease_data, col_dates, col_cases, top = 3, concat_values = T)
}
