% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_months_most_cases}
\alias{get_months_most_cases}
\title{Get months major cases}
\usage{
get_months_most_cases(
  disease_data,
  col_dates,
  col_cases = "casos",
  top = 3,
  concat_values = TRUE
)
}
\arguments{
\item{disease_data}{The disease data}

\item{col_dates}{Column names in the disease data that contains the dates}

\item{col_cases}{Column name in the disease data that contains
the cases number}

\item{top}{Top}

\item{concat_values}{Indicates if it is required to concatenate
the months as a string}
}
\value{
Months major cases
}
\description{
Function that gets the months with the major number of cases
}
\examples{
disease_data <- import_linelist_disease_year(2020, "DENGUE")
disease_data <- clean_header(disease_data)
cases_onset_symptoms <- group_onset_symptoms(disease_data, type = "day")
get_months_most_cases(cases_onset_symptoms,
                       col_dates = "ini_sin",
                       col_cases = "casos",
                       top = 3,
                       concat_values = TRUE)
}
