% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_most_cases}
\alias{get_most_cases}
\title{Get row with major cases}
\usage{
get_most_cases(disease_data, col_name = "casos", percentage = TRUE)
}
\arguments{
\item{disease_data}{The disease data}

\item{col_name}{Column name in the disease data that contains the cases number}

\item{percentage}{Indicates if it is required to add a percentage of cases as a column}
}
\value{
Row with major cases
}
\description{
Function that gets the row with the major number of cases
}
\examples{
disease_data <- import_data_disease_by_year(2020, "DENGUE")
get_most_cases(disease_data, col_name = "casos", percentage = T)
}
