% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{get_name_file_path}
\alias{get_name_file_path}
\title{Get Name File Path or URL}
\usage{
get_name_file_path(path)
}
\arguments{
\item{path}{The path or URL}
}
\value{
The name file
}
\description{
Función que obtiene el nombre del archivo a descargar desde una URL o ruta\n
Function that gets the file name for download from a URL or path
}
\examples{
find_name_file_path("DENGUE")
}
