% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_name_mpios}
\alias{get_name_mpios}
\title{Get name of municipality in Colombia}
\usage{
get_name_mpios(geo_data, code_dept, code_mpio)
}
\arguments{
\item{code_dept}{The department code}

\item{code_mpio}{The municipality code}
}
\value{
The municipality name
}
\description{
Function that gets the municipality name
}
\examples{
get_departments()
}
