% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_name_muns}
\alias{get_name_muns}
\title{Get name of municipality in Colombia}
\usage{
get_name_muns(geo_data, code_dept, code_mun)
}
\arguments{
\item{code_dept}{The department code}

\item{code_mun}{The municipality code}
}
\value{
The municipality name
}
\description{
Function that gets the municipality name
}
\examples{
geo_data <- import_geo_codes()
get_name_muns(geo_data,
              code_dept = "05",
              code_mun = "001")
}
