% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping_data.R
\name{get_path_data_disease_by_year}
\alias{get_path_data_disease_by_year}
\title{Get Path Data Disease by Year}
\usage{
get_path_data_disease_by_year(year, disease_name)
}
\arguments{
\item{year}{The year}

\item{disease_name}{The disease name}
}
\value{
Path or URL for download the disease data by selected year
}
\description{
Función que obtiene la ruta o URL del servicio para descargar la informacion de una enfermedad por año desde SIVIGILA
Function that obtains the path or URL of the service to download the information of a disease per year from SIVIGILA
}
\examples{
get_path_data_disease_by_year(2010, "DENGUE")
}
