% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{get_special_population_and_cases}
\alias{get_special_population_and_cases}
\title{Get Special Population and Cases}
\usage{
get_special_population_and_cases(disease_data)
}
\arguments{
\item{disease_data}{Disease data}
}
\value{
The cases by special population type of a disease
}
\description{
Función que obtiene los casos por tipo de población especial de una enfermedad
Function that gets the cases by special population type of a disease
}
\examples{
get_special_population_and_cases(disease_data)
}
