% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{get_special_population_cases}
\alias{get_special_population_cases}
\title{Get special population and cases}
\usage{
get_special_population_cases(disease_data)
}
\arguments{
\item{disease_data}{The disease data}
}
\value{
The cases by special population type of a disease
}
\description{
Function that gets the cases by special population type of a disease
}
\examples{
disease_data <- import_linelist_disease_year(2019, "DENGUE")
get_special_population_cases(disease_data)
}
