% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_age}
\alias{group_age}
\title{Group by age and cases}
\usage{
group_age(
  disease_data,
  col_name = "edad",
  percentage = FALSE,
  age_interval = 10
)
}
\arguments{
\item{disease_data}{The disease data}

\item{col_name}{Column name in the disease data that contains the ages}

\item{percentage}{Indicates if it is required to add a percentage of cases as a column}

\item{age_interval}{The interval of age range}
}
\value{
The disease data grouped by age and cases
}
\description{
Function that groups the disease data by age and cases
}
\examples{
disease_data <- import_data_disease_by_year(2019, "DENGUE")
group_age(disease_data, col_name = "edad", percentage = FALSE)
}
