% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_age_epiweek}
\alias{group_age_epiweek}
\title{Group by age, epidemiological week and cases}
\usage{
group_age_epiweek(
  disease_data,
  col_names = c("edad", "semana"),
  percentage = FALSE
)
}
\arguments{
\item{disease_data}{The disease data}

\item{col_names}{Column names in the disease data that contains the ages and the epidemiological weeks}

\item{percentage}{Indicates if it is required to add a percentage of cases as a column}
}
\value{
The disease data grouped by age, epiweek and cases
}
\description{
Function that groups the disease data by age, epidemiological week and cases
}
\examples{
disease_data <- import_data_disease_by_year(2019, "DENGUE")
group_age_epiweek(disease_data, col_names = c("edad", "semana"), percentage = FALSE)
}
