% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_age_sex}
\alias{group_age_sex}
\title{Group by ages, sex and cases}
\usage{
group_age_sex(
  disease_data,
  col_names = c("edad", "sexo"),
  percentage = TRUE,
  age_interval = 10
)
}
\arguments{
\item{disease_data}{The disease data}

\item{col_names}{Column names in the disease data that contains
the ages and sex}

\item{percentage}{Indicates if it is required to add a
percentage of cases as a column}

\item{age_interval}{The interval of age range}
}
\value{
The disease data grouped by ages, sex and number cases
}
\description{
Function that groups the data by age, sex and cases number
}
\examples{
disease_data <- import_linelist_disease_year(2019, "DENGUE")
disease_data <- clean_header(disease_data)
group_age_sex(disease_data, col_names = c("edad", "sexo"), percentage = TRUE)
}
