% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_by_age_and_gender}
\alias{group_by_age_and_gender}
\title{group_by_age_and_gender}
\usage{
group_by_age_and_gender(
  disease_data,
  col_names = c("edad", "sexo"),
  percentage = T
)
}
\arguments{
\item{disease_data}{Disease data}

\item{col_names}{Data set column names}
}
\value{
The disease data grouped by age, gender and cases
}
\description{
Función que agrupa los datos por edad, genero y los casos
Function that groups the data by age, gender and cases
}
\examples{
disease_data <- import_data_disease_by_year(2019, "DENGUE")
group_by_age_and_gender(disease_data, col_names = c("edad", "sexo"), percentage = T)
}
