% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_by_age_range_and_cases}
\alias{group_by_age_range_and_cases}
\title{Group by age range and cases}
\usage{
group_by_age_range_and_cases(
  disease_data,
  col_name,
  var_a = NULL,
  min_val,
  max_val,
  step
)
}
\arguments{
\item{disease_data}{The disease data}

\item{col_name}{Column name in the disease data that contains the ages}

\item{min_val}{Minimum value of ages}

\item{max_val}{Maximum value of ages}

\item{step}{Step to generate the age range}
}
\value{
The disease data grouped by the age range and cases number
}
\description{
Function that groups the disease data by age range and cases number
}
\examples{
disease_data <- import_data_disease_by_year(2019, "DENGUE")
disease_dt_by_age <- group_by_columns_and_cases(disease_data, c("edad", "semana"),
  wt_percentage = TRUE
)
group_by_age_range_and_cases(disease_dt_by_age, "edad",
  min_val = 0,
  max_val = max(disease_dt_by_age$edad), step = 10
)
}
