% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_by_deptos}
\alias{group_by_deptos}
\title{group_by_deptos}
\usage{
group_by_deptos(disease_data, col_name = "cod_dpto_o", percentage = F)
}
\arguments{
\item{disease_data}{Disease data}

\item{col_name}{Data set column name}
}
\value{
The disease data grouped by code departments and cases
}
\description{
Función que agrupa los datos por codigo de departamento y casos
Function that groups the data by code departments and cases
}
\examples{
disease_data <- import_data_disease_by_year(2019, "DENGUE")
group_by_codes_deptos(disease_data, col_name = "cod_dpto_o", percentage = F)
}
