% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_by_gender_and_week}
\alias{group_by_gender_and_week}
\title{group_by_gender_and_week}
\usage{
group_by_gender_and_week(
  disease_data,
  col_names = c("sexo", "semana"),
  percentage = T
)
}
\arguments{
\item{disease_data}{Disease data}

\item{col_names}{Data set column names}
}
\value{
The disease data grouped by gender, epiweek and cases
}
\description{
Función que agrupa los datos por genero, semana epidemiologica y los casos
Function that groups the data by gender, epiweek and cases
}
\examples{
disease_data <- import_data_disease_by_year(2019, "DENGUE")
group_by_columns_and_cases(disease_data, col_name = "sexo",)
}
