% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_by_notification_date}
\alias{group_by_notification_date}
\title{group_by_notification_date}
\usage{
group_by_notification_date(disease_data, col_name = "fec_not", type = "month")
}
\arguments{
\item{disease_data}{Disease data}

\item{col_name}{Data set column name}

\item{type}{Time unit}
}
\value{
The disease data grouped by onset symptoms date and cases
}
\description{
Función que agrupa los datos por la fecha de notificacion y los casos
Function that groups the data by notification date and cases
}
\examples{
disease_data <- import_data_disease_by_year(2019, "DENGUE")
group_by_notification_date(disease_data, col_name = "fec_not", type = "month")
}
