% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_by_range}
\alias{group_by_range}
\title{Group By Range}
\usage{
group_by_range(disease_data, var, var_a = NULL, min_val, max_val, step)
}
\arguments{
\item{disease_data}{The disease data}

\item{var}{The variable to group}

\item{min_val}{The min value to group}

\item{max_val}{The max value to group}

\item{step}{The step for the range}
}
\value{
The disease data grouped by the range
}
\description{
Función que agrupa los datos por un rango
Function that groups the data by a range
}
\examples{
group_by_range(disease_data, min_val = 2020-10-10, max_val = 2022-12-10, step = 5)
}
