% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_by_special_population}
\alias{group_by_special_population}
\title{group_by_special_population}
\usage{
group_by_special_population(
  disease_data,
  col_name = "poblacion",
  percentage = T
)
}
\arguments{
\item{disease_data}{Disease data}

\item{col_name}{Data set column name}
}
\value{
The disease data grouped by  special population and cases
}
\description{
Función que agrupa los datos por poblacion especial y casos
Function that groups the data by special population and cases
}
\examples{
disease_data <- import_data_disease_by_year(2019, "DENGUE")
group_by_special_population(disease_data, col_name = "poblacion", percentage = T)
}
