% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_by_week}
\alias{group_by_week}
\title{Group By Week}
\usage{
group_by_week(disease_data)
}
\arguments{
\item{disease_data}{Disease data}
}
\value{
The disease data grouped by week
}
\description{
Función que agrupa los datos por Semana
Function that groups the data by week
}
\examples{
group_by_week(disease_data)
}
