% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_by_week_and_cases}
\alias{group_by_week_and_cases}
\title{Group By Week and Cases}
\usage{
group_by_week_and_cases(disease_data)
}
\arguments{
\item{disease_data}{Disease data}
}
\value{
The disease data grouped by week
}
\description{
Función que agrupa los datos por semana epidemiológica y número de casos
Function that groups the data by epidemiological week and number of cases
}
\examples{
group_by_week_and_cases(disease_data)
}
