% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_dept}
\alias{group_dept}
\title{Group by department and cases}
\usage{
group_dept(disease_data, col_name = "cod_dpto_o", percentage = FALSE)
}
\arguments{
\item{disease_data}{The disease data}

\item{col_name}{Column name in the disease data that contains 
the department codes}

\item{percentage}{Indicates if it is required to add a 
percentage of cases as a column}
}
\value{
The disease data grouped by department codes and cases number
}
\description{
Function that groups the data by code departments and cases number
}
\examples{
disease_data <- import_linelist_disease_year(2019, "DENGUE")
group_dept(disease_data, col_name = "cod_dpto_o", percentage = FALSE)
}
