% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_mun}
\alias{group_mun}
\title{Group by municipalities and cases}
\usage{
group_mun(
  disease_data,
  dept_name = NULL,
  col_name = "cod_mun_o",
  percentage = FALSE
)
}
\arguments{
\item{disease_data}{The disease data}

\item{dept_name}{The department name}

\item{col_name}{Column name in the disease data that contains
the municipalities codes}

\item{percentage}{Indicates if it is required to add a
percentage of cases as a column}
}
\value{
The disease data grouped by municipality codes and cases number
}
\description{
Function that groups the data by code municipalities and cases number
}
\examples{
disease_data <- import_linelist_disease_year(2019, "DENGUE")
disease_data <- clean_header(disease_data)
group_mun(disease_data,
           dept_name = "Antioquia",
           col_name = "cod_mun_o",
           percentage = FALSE)
}
