% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_municipalities}
\alias{group_municipalities}
\title{Group by municipalities and cases}
\usage{
group_municipalities(
  disease_data,
  department = NULL,
  col_name = "cod_mun_o",
  percentage = FALSE
)
}
\arguments{
\item{disease_data}{The disease data}

\item{col_name}{Column name in the disease data that contains
the municipalities codes}

\item{percentage}{Indicates if it is required to add a
percentage of cases as a column}
}
\value{
The disease data grouped by municipality codes and cases number
}
\description{
Function that groups the data by code municipalities and cases number
}
\examples{
disease_data <- import_linelist_disease_year(2019, "DENGUE")
disease_data <- clean_header(disease_data)
group_municipalities(disease_data, col_name = "cod_mun_o", percentage = FALSE)
}
