% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_notification_date}
\alias{group_notification_date}
\title{Group by notification date and cases}
\usage{
group_notification_date(disease_data, col_name = "fec_not", type = "month")
}
\arguments{
\item{disease_data}{The disease data}

\item{col_name}{Column name in the disease data that contains 
the notification dates}

\item{type}{Time unit (day, month and year)}
}
\value{
The disease data grouped by notification date and cases
}
\description{
Function that groups the disease data by notification date and cases number
}
\examples{
disease_data <- import_linelist_disease_year(2019, "DENGUE")
disease_data <- clean_header(disease_data)
group_notification_date(disease_data, col_name = "fec_not", type = "month")
}
