% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_onset_symptoms}
\alias{group_onset_symptoms}
\title{Group by symptom onset date and cases}
\usage{
group_onset_symptoms(disease_data, col_name = "ini_sin", type = "month")
}
\arguments{
\item{disease_data}{The disease data}

\item{col_name}{Column name in the disease data that contains
the symptom onset dates}

\item{type}{Time unit (day, month and year)}
}
\value{
The disease data grouped by symptom onset date and number of cases
}
\description{
Function that groups the disease data by symptom onset date and
number of cases
}
\examples{
disease_data <- import_linelist_disease_year(2019, "DENGUE")
disease_data <- clean_header(disease_data)
group_onset_symptoms(disease_data, col_name = "ini_sin", type = "month")
}
