% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_sex}
\alias{group_sex}
\title{Group by sex and cases}
\usage{
group_sex(disease_data, col_name = "sexo", percentage = TRUE)
}
\arguments{
\item{disease_data}{The disease data}

\item{col_name}{Column name in the disease data that contains the sex}

\item{percentage}{Indicates if it is required to add a 
percentage of cases as a column}
}
\value{
The disease data grouped by sex and cases number
}
\description{
Function that groups the disease data by sex and cases number
}
\examples{
disease_data <- import_linelist_disease_year(2019, "DENGUE")
disease_data <- clean_header(disease_data)
group_sex(disease_data, col_name = "sexo", percentage = TRUE)
}
