% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_special_population}
\alias{group_special_population}
\title{Group by special population and cases}
\usage{
group_special_population(
  disease_data,
  col_name = "poblacion",
  percentage = TRUE
)
}
\arguments{
\item{disease_data}{The disease data}

\item{col_name}{Column name in the disease data that contains
the special populations}

\item{percentage}{Indicates if it is required to add a
percentage of cases as a column}
}
\value{
The disease data grouped by special populations and cases
}
\description{
Function that groups the disease data by special population and cases
}
\examples{
disease_data <- import_linelist_disease_year(2019, "DENGUE")
disease_data <- clean_header(disease_data)
group_special_population(disease_data,
                         col_name = "poblacion",
                         percentage = TRUE)
}
