% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data_canal_endemico}
\alias{import_data_canal_endemico}
\title{Importar datos de una enfermedad para construir su canal endémico}
\usage{
import_data_canal_endemico(nombre_event, year)
}
\arguments{
\item{nombre_event}{El nombre del evento}

\item{year}{Año final}
}
\value{
Los datos de los últimos cinco años de una enfermedad
}
\description{
Función que importa los últimos cinco años de datos de
una enfermedad para construir
el canal endémico desde la fuente de SIVIGILA
}
\examples{
import_data_canal_endemico("MALARIA", 2020)
}
