% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data_delim}
\alias{import_data_delim}
\title{Import Data Delim}
\usage{
import_data_delim(path_data)
}
\arguments{
\item{path_data}{Path of SIVIGILA data}
}
\value{
Data tabulated
}
\description{
Función que importa la informacion del SIVIGILA y la tabula a traves de un delimitador
Function that imports the SIVIGILA information and tabulates it through a delimiter
}
\examples{
import_data_delim(https://www.datos.gov.co/api/views/qvnt-2igj/rows.csv?accessType=DOWNLOAD)
}
