% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data_disease_by_year}
\alias{import_data_disease_by_year}
\title{Import Data of a Disease By Year}
\usage{
import_data_disease_by_year(year, disease_name, cache = TRUE)
}
\arguments{
\item{year}{The year}

\item{disease_name}{The disease name}
}
\value{
The disease data by year
}
\description{
Función que obtiene los datos de la enfermedad por año
Function that obtains the disease data by year
}
\examples{
import_data_disease_by_year(2018, "DENGUE")
}
