% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data_endemic_channel}
\alias{import_data_endemic_channel}
\title{Import Data Endemic Channel}
\usage{
import_data_endemic_channel(disease_name, year)
}
\arguments{
\item{disease_name}{The disease name}

\item{year}{Last year}

\item{path_data}{Path or URL of SIVIGILA data}
}
\value{
The last five years data of a disease
}
\description{
Función que importa la informacion del SIVIGILA para la construcción del canal endémico
Function that imports SIVIGILA for building the endemic channel
}
\examples{
import_data_endemic_channel("MALARIA", 2020)
}
