% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data_event}
\alias{import_data_event}
\title{Importar datos de una enfermedad por año desde los microdatos de SIVIGILA}
\usage{
import_data_event(year, nombre_event, cache = TRUE)
}
\arguments{
\item{year}{El año seleccionado}

\item{nombre_event}{El nombre de la enfermedad}

\item{cache}{Indica si los datos descargados deben ser almacenados en caché}
}
\value{
Los datos de la enfermedad por año desde los microdatos de SIVIGILA
}
\description{
Función que obtiene los datos de una enfermedad por año
}
\examples{
import_data_event(2018, "DENGUE")
}
