% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data_incidencia}
\alias{import_data_incidencia}
\title{Importar las proyecciones DANE del año 2005 hasta el 2035}
\usage{
import_data_incidencia()
}
\value{
Un `data.frame` con las proyecciones poblacionales
}
\description{
Función que obtiene las proyecciones poblacionales DANE desde
el año 2005 hasta el 2035
}
\examples{
import_data_incidencia()
}
