% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data_resumen_sivigila}
\alias{import_data_resumen_sivigila}
\title{Importar datos resumidos de SIVIGILA}
\usage{
import_data_resumen_sivigila(url_data = NULL)
}
\arguments{
\item{url_data}{URL de los datos de SIVIGILA}
}
\value{
Los datos descargados en formato csv
}
\description{
Función que importa datos de SIVIGILA a través de una URL
}
\examples{
import_data_resumen_sivigila()
}
