% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data_separator}
\alias{import_data_separator}
\title{Import data with a specific separator}
\usage{
import_data_separator(path_data)
}
\arguments{
\item{path_data}{Path or URL of SIVIGILA data}
}
\value{
A data frame
}
\description{
Function that identifies the separator with which the information comes to be able to tabulate it
}
\examples{
import_data_separator("https://www.datos.gov.co/api/views/qvnt-2igj/rows.csv?accessType=DOWNLOAD")
}
