% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_deptos_data}
\alias{import_deptos_data}
\title{Import Departments Data}
\usage{
import_deptos_data(url_data = config::get("deptos_data_path"))
}
\arguments{
\item{url_data}{URL of departments data}
}
\value{
The data downloaded in csv format
}
\description{
Función que importa la informacion de los departamentos de Colombia a través de una URL
Function that imports the information of the departments of Colombia through a URL
}
\examples{
import_deptos_data(""https://www.datos.gov.co/api/views/gdxc-w37w/rows.csv?accessType=DOWNLOAD"")
}
