% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_geo_codes}
\alias{import_geo_codes}
\title{Import geographical data of Colombia}
\usage{
import_geo_codes(url_data = NULL)
}
\arguments{
\item{url_data}{URL of geographical data}
}
\value{
A data frame with the names and codes of the departments and
municipalities of Colombia in csv format
}
\description{
Function that imports the names and codes of the departments and
municipalities of Colombia through a URL
}
\examples{
import_geo_codes(
 "https://www.datos.gov.co/api/views/gdxc-w37w/rows.csv?accessType=DOWNLOAD")
}
