% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_geo_cods}
\alias{import_geo_cods}
\title{Importar datos geográficos de Colombia}
\usage{
import_geo_cods(url_data = NULL)
}
\arguments{
\item{url_data}{URL de los datos geográficos}
}
\value{
Un data frame con los nombres y códigos de los departamentos
y municipios de Colombia en formato csv
}
\description{
Función que importa los nombres y códigos de los departamentos
y municipios de Colombia a través de una URL
}
\examples{
import_geo_cods(
"https://www.datos.gov.co/api/views/gdxc-w37w/rows.csv?accessType=DOWNLOAD")
}
