% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_path_data}
\alias{import_path_data}
\title{Import Path Data}
\usage{
import_path_data(url_data, cache = TRUE)
}
\arguments{
\item{url_data}{URL of SIVIGILA data}
}
\value{
The data downloaded in csv format
}
\description{
Función que importa la informacion de SIVIGILA a través de una URL
Function that imports SIVIGILA data through a URL
}
\examples{
import_path_data("https://www.datos.gov.co/api/views/qvnt-2igj/rows.csv?accessType=DOWNLOAD")
}
