% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{limpiar_data_incidencia}
\alias{limpiar_data_incidencia}
\title{Limpiar datos o proyecciones DANE}
\usage{
limpiar_data_incidencia(proyecciones)
}
\arguments{
\item{proyecciones}{Un `data.frame` que contiene los datos de
las proyecciones DANE}
}
\value{
Un `data.frame` con los datos limpios y estandarizados de las
proyecciones DANE
}
\description{
Función que limpia y estandariza las proyecciones DANE para realizar
el cálculo de incidencia
}
\examples{
proyecciones = import_data_incidencia()
limpiar_data_sivigila(proyecciones = proyecciones)
}
