% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{limpiar_data_sivigila}
\alias{limpiar_data_sivigila}
\title{Limpiar datos de SIVIGILA}
\usage{
limpiar_data_sivigila(data_event, uni_med_edad = 1)
}
\arguments{
\item{data_event}{Un `data.frame` que contiene los datos de
una enfermedad o evento.}

\item{uni_med_edad}{Un `numeric` (numérico) o `character`(cadena de
caracteres) que contiene la unidad de medida a la que se debe estandarizar
la edad; su valor por defecto es `1`.}
}
\value{
Un `data.frame` con los datos limpios de la enfermedad o evento.
}
\description{
Función que limpia los datos seleccionados de una enfermedad
o evento provenientes de la fuente SIVIGILA.
}
\examples{
data(dengue2020)
limpiar_data_sivigila(data_event = dengue2020)
}
