% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{limpiar_fecha_event}
\alias{limpiar_fecha_event}
\title{Limpiar fechas de los datos de una enfermedad o evento}
\usage{
limpiar_fecha_event(
  data_event,
  year,
  format_fecha = "\%AAAA-\%MM-\%DD",
  nombre_col = "ini_sin",
  col_comp = NULL
)
}
\arguments{
\item{data_event}{Los datos de una enfermedad o evento}

\item{year}{Año de los datos de una enfermedad o evento}

\item{format_fecha}{Formato de fecha}

\item{nombre_col}{Nombre de la columna del conjunto de datos}

\item{col_comp}{Nombre de la columna de comparación del conjunto de datos}
}
\value{
Los datos con las fechas limpias
}
\description{
Función que limpia las fechas de los datos de una enfermedad o evento
}
\examples{
data_event <- import_data_event(2020, "DENGUE")
data_event <- limpiar_encabezado(data_event)
limpiar_fecha_event(data_event,
                    year = 2020,
                    format_fecha = "\%AAAA-\%MM-\%DD",
                    nombre_col = "ini_sin",
                    col_comp = "fec_hos")
}
