% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{list_available_diseases_and_years}
\alias{list_available_diseases_and_years}
\title{Import diseases and years available for download}
\usage{
list_available_diseases_and_years()
}
\value{
The diseases and the years available from the SIVIGILA microdata source
}
\description{
Function that obtains the diseases and the years available from the SIVIGILA microdata source
}
\examples{
list_available_diseases_and_years()
}
