% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{list_avaliable_diseases_and_years}
\alias{list_avaliable_diseases_and_years}
\title{Import Avaliable Diseases and Years}
\usage{
list_avaliable_diseases_and_years()
}
\value{
The diseases and the years available
}
\description{
Función que obtiene las enfermedades y los años disponibles de los microdatos de SIVIGILA
Function that obtains the diseases and the years available from the SIVIGILA microdata
}
\examples{
import_avaliable_diseases_and_years()
}
