% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{listar_dptos}
\alias{listar_dptos}
\title{Listar códigos de departamentos}
\usage{
listar_dptos(geo_cods)
}
\arguments{
\item{geo_cods}{Un `data.frame` que contiene la información geográfica
(departamentos y municipios de Colombia) con sus códigos}
}
\value{
Un `data.frame` con los datos de los departamentos con
código y nombre
}
\description{
Función que lista los departamentos de Colombia con su nombre y
código
}
\examples{
geo_cods <- import_geo_cods()
listar_dptos(geo_cods = geo_cods)
}
