% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modficar_cod_mun}
\alias{modficar_cod_mun}
\title{Establecer códigos geográficos de los datos de la enfermedad}
\usage{
modficar_cod_mun(code_dept, cod_mun)
}
\arguments{
\item{code_dept}{El código del departamento}

\item{cod_mun}{El código del municipio}
}
\value{
Los códigos geográficos de los datos de la enfermedad
}
\description{
Función que establece los códigos geográficos de los datos de la enfermedad
}
\examples{
modficar_cod_mun(code_dept = 01, cod_mun = "001")
}
