% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{obtener_casos_pob_especial}
\alias{obtener_casos_pob_especial}
\title{Obtener poblaci贸n especial y casos}
\usage{
obtener_casos_pob_especial(data_event)
}
\arguments{
\item{data_event}{Un data frame que contiene los datos de una
enfermedad o evento}
}
\value{
Un data frame con los casos por tipo de poblaci贸n especial
de una enfermedad o evento
}
\description{
Funci贸n que obtiene los casos por tipo de poblaci贸n
especial de una enfermedad
}
\examples{
data(dengue2020)
data_event <- dengue2020
data_event <- limpiar_data_sivigila(data_event, 2020)
obtener_casos_pob_especial(data_event = data_event)
}
