% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{obtener_cods_dpto}
\alias{obtener_cods_dpto}
\title{Obtener códigos de departamentos}
\usage{
obtener_cods_dpto(geo_cods)
}
\arguments{
\item{geo_cods}{Códigos geográficos (departamentos y municipios de
Colombia)}
}
\value{
Los datos de los departamentos con código y nombre
}
\description{
Función que obtiene la lista de departamentos de Colombia con su nombre
y código
}
\examples{
geo_cods <- import_geo_cods()
obtener_cods_dpto(geo_cods)
}
