% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtener_cond_inciden_event}
\alias{obtener_cond_inciden_event}
\title{Obtener las condiciones para calcular la incidencia de una
enfermedad o evento}
\usage{
obtener_cond_inciden_event(cod_eve)
}
\arguments{
\item{cod_eve}{Un `numeric` (númerico) o `character` (cadena de
caracteres) que contiene el código de una enfermedad o evento.}
}
\value{
Un `data.frame` con las condiciones para calcular la
incidencia de una enfermedad o evento.
}
\description{
Función que obtiene las condiciones del numerador,
denominador y coeficiente de múltiplicación para calcular la incidencia
de un evento.
}
\examples{
obtener_cond_inciden_event(cod_eve = 210)
}
