% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtener_config_map}
\alias{obtener_config_map}
\title{Obtener la configuración del mapa}
\usage{
obtener_config_map(data_agrupada, dpto, mpio, geo_ocurrencia, shp)
}
\arguments{
\item{data_agrupada}{Un `data.frame` que contiene los datos de la enfermedad
agrupados por departamento y número de casos.}

\item{dpto}{Un `character` (cadena de caracteres) que contiene el
nombre del departamento.}

\item{mpio}{Un `character` (cadena de caracteres) que contiene el
nombre del municipio.}

\item{geo_ocurrencia}{Un `data.frame` con las columnas de ocurrencia
geográfica de los datos de la enfermedad o evento.}

\item{shp}{Objeto que contiene el`Shapefile` del mapa.}
}
\value{
Una `named list` (lista nombrada) que contiene el departamento,
municipio y el poligono para generar el mapa con los siguientes
nombres para cada uno de estos elementos: `dpto`, `mpio` y `poligono`.
}
\description{
Función que obtiene el departamento, municipio y
el poligono requeridos para generar el mapa.
}
\keyword{internal}
