% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_sivirep.R
\name{obtener_estetica_escala}
\alias{obtener_estetica_escala}
\title{Obtener la estetica de una escala para un grafico de \pkg{sivirep}
Función que genera la estetica de una escala para un grafico de
\pkg{sivirep}}
\usage{
obtener_estetica_escala(escala = 0, nombre)
}
\arguments{
\item{escala}{Un numero (numeric) que indica la cantidad de valores que
contiene la escala}

\item{nombre}{Un character (cadena de caracteres) que contiene el nombre
de la escala}
}
\value{
Un objecto scale_fill_manual de \pkg{ggplot2}
}
\description{
Obtener la estetica de una escala para un grafico de \pkg{sivirep}
Función que genera la estetica de una escala para un grafico de
\pkg{sivirep}
}
\examples{
library(ggplot2)

ggplot(mtcars, ggplot2::aes(wt, mpg)) +
  ggplot2::geom_point() +
  ggplot2::labs(title = "La economia") +
  tema_sivirep()
}
