% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtener_fila_mas_casos}
\alias{obtener_fila_mas_casos}
\title{Obtener fila con mayor número de casos}
\usage{
obtener_fila_mas_casos(data_event, col_name = "casos", porcentaje = TRUE)
}
\arguments{
\item{data_event}{Los datos de la enfermedad}

\item{col_name}{Nombre de la columna en los datos de la
enfermedad o evento que contiene el número de casos}

\item{porcentaje}{Indica si se requiere agregar un porcentaje
de casos como columna}
}
\value{
La fila con mayor número de casos
}
\description{
Función que obtiene la fila con el mayor número de casos
}
\examples{
data_event <- import_data_event(2020, "DENGUE")
data_event <- limpiar_encabezado(data_event)
casos_sex <- agrupar_sex(data_event,
                         porcentaje = TRUE)
obtener_fila_mas_casos(casos_sex,
                       col_name = "casos",
                       porcentaje = TRUE)
}
