% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtener_info_depts}
\alias{obtener_info_depts}
\title{Obtener información geográfica de los datos de la enfermedad}
\usage{
obtener_info_depts(department = NULL, municip = NULL)
}
\arguments{
\item{department}{El nombre del departamento}

\item{municip}{Los datos del municipio}
}
\value{
La información geográfica de los datos de la enfermedad
}
\description{
Función que obtiene la información geográfica de los datos de la enfermedad
}
\examples{
obtener_info_depts(department = "ANTIOQUIA")
}
