% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtener_meses_mas_casos}
\alias{obtener_meses_mas_casos}
\title{Obtener los meses con mayor número de casos}
\usage{
obtener_meses_mas_casos(
  data_event,
  col_fechas,
  col_casos = "casos",
  top = 3,
  concat_vals = TRUE
)
}
\arguments{
\item{data_event}{Los datos de la enfermedad}

\item{col_fechas}{Nombres de columna en los datos de la
enfermedad o evento que contienen las fechas}

\item{col_casos}{Nombre de la columna en los datos de la
enfermedad o evento que contiene
el número de casos}

\item{top}{Cantidad máxima de meses a retornar}

\item{concat_vals}{Indica si se requiere concatenar los meses
como una cadena}
}
\value{
Los meses con mayor número de casos
}
\description{
Función que obtiene los meses con el mayor número de casos
}
\examples{
data_event <- import_data_event(2020, "DENGUE")
data_event <- limpiar_encabezado(data_event)
casos_inisintomas <- agrupar_fecha_inisintomas(data_event, tipo = "day")
obtener_meses_mas_casos(casos_inisintomas,
                    col_fechas = "ini_sin",
                    col_casos = "casos",
                    top = 3,
                    concat_vals = TRUE)
}
