% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtener_nombres_dptos}
\alias{obtener_nombres_dptos}
\title{Obtener nombres de departamentos}
\usage{
obtener_nombres_dptos(data_event)
}
\arguments{
\item{data_event}{Un data frame que contiene los datos de
la enfermedad o evento}
}
\value{
Un data frame con los nombres de los departamentos
}
\description{
Función que obtiene los nombres de los departamentos
}
\examples{
data(dengue2020)
data_event <- dengue2020
}
