% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtener_nombres_muns}
\alias{obtener_nombres_muns}
\title{Obtener nombre del municipio en Colombia}
\usage{
obtener_nombres_muns(data_geo, code_dept, cod_mun)
}
\arguments{
\item{data_geo}{Códigos geográficos (departamentos y municipios de
Colombia)}

\item{code_dept}{El código del departamento}

\item{cod_mun}{El código del municipio}
}
\value{
El nombre del municipio
}
\description{
Función que obtiene el nombre del municipio
}
\examples{
data_geo <- import_geo_cods()
obtener_nombres_muns(data_geo,
                     code_dept = "05",
                     cod_mun = "001")
}
