% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping_data.R
\name{obtener_ruta_data_event_year}
\alias{obtener_ruta_data_event_year}
\title{Obtener la ruta de descarga de una enfermedad por un año específico}
\usage{
obtener_ruta_data_event_year(nombre_event, year)
}
\arguments{
\item{nombre_event}{Un `character` (cadena de caracteres) con el nombre de
la enfermedad o evento.}

\item{years}{Un `numeric` (numerico) con el año o años deseado(s) para
la descarga de los datos.}
}
\value{
Un `character` (cadena de caracteres) con la ruta o URL para
descargar los datos de una enfermedad o evento por un año específico de la
fuente SIVIGILA.
}
\description{
Función que obtiene la ruta o URL del servicio SIVIGILA para
descargar los datos de una enfermedad o evento para un año específico.
}
\keyword{internal}
