% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{obtener_ruta_descarga}
\alias{obtener_ruta_descarga}
\title{Obtener el nombre del archivo desde una URL}
\usage{
obtener_ruta_descarga(ruta)
}
\arguments{
\item{ruta}{Un character (cadena de caracteres) con la ruta o URL
de descarga}
}
\value{
Un character (cadena de caracteres) con el nombre del
archivo a descargar
}
\description{
Función que obtiene el nombre del archivo a descargar desde una URL o ruta
}
\examples{
obtener_ruta_descarga("DENGUE")
}
