% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtener_val_config}
\alias{obtener_val_config}
\title{Obtener valor del archivo de configuración}
\usage{
obtener_val_config(llave)
}
\arguments{
\item{llave}{Un `character` (cadena de caracteres) con el nombre de la
llave que se encuentra en el archivo de configuración del paquete.}
}
\value{
Un `character` (cadena de caracteres) con el valor de la llave
del archivo de configuración del paquete.
}
\description{
Función que obtiene el valor de una llave del archivo de
configuración.
}
\examples{
obtener_val_config("request_timeout")
}
