% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtener_year}
\alias{obtener_year}
\title{Obtener el año de una enfermedad o evento}
\usage{
obtener_year(data_event)
}
\arguments{
\item{data_event}{Un `data.frame` que contiene los datos de la
enfermedad o evento.}
}
\value{
Un `numeric` (numérico) que representa el año de los datos de la
enfermedad o evento.
}
\description{
Función que obtiene el año de los datos de una enfermedad o
evento.
}
\keyword{internal}
