% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_age_and_gender}
\alias{plot_age_and_gender}
\title{plot_age_and_gender}
\usage{
plot_age_and_gender(
  disease_data,
  col_names = c("edad", "sexo"),
  percentage = T
)
}
\arguments{
\item{disease_data}{Disease data}

\item{col_names}{Data set column names}
}
\value{
A plot of cases distribution by age and gender
}
\description{
Función que genera la gráficas de distribucion de casos por edad y genero
Function that generates the plot of cases distribution by age and gender
}
\examples{
disease_data <- import_data_disease_by_year(2020, "DENGUE")
plot_age_and_gender(disease_data, col_names = c("edad", "sexo"), percentage = T)
}
