% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_age_gender}
\alias{plot_age_gender}
\title{Plot cases distribution by age and gender}
\usage{
plot_age_gender(
  disease_data,
  col_names = c("edad", "sexo"),
  percentage = FALSE
)
}
\arguments{
\item{disease_data}{The disease data}

\item{col_names}{Column names in the disease data that contains the ages and the epidemiological weeks}
}
\value{
A plot of cases distribution by age and gender
}
\description{
Function that generates the plot of cases distribution by age and gender
}
\examples{
disease_data <- import_data_disease_by_year(2020, "DENGUE")
disease_data <- group_by_age_and_week(disease_data, col_names = c("edad", "semana"), percentage = FALSE)
plot_age_gender(disease_data, col_names = c("edad", "sexo"), percentage = FALSE)
}
