% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_age_sex}
\alias{plot_age_sex}
\title{Plot cases distribution by age and sex}
\usage{
plot_age_sex(data_grouped, col_names = c("edad", "sexo"), percentage = FALSE)
}
\arguments{
\item{data_grouped}{The disease data grouped}

\item{col_names}{Column names in the disease data grouped that contains}

\item{percentage}{Indicates if the data has percentages
the ages and the epidemiological weeks}
}
\value{
A plot of cases distribution by age and sex
}
\description{
Function that generates the plot of cases distribution by age and sex
}
\examples{
disease_data <- import_linelist_disease_year(2020, "DENGUE")
disease_data <- clean_header(disease_data)
data_grouped <- group_age_sex(disease_data,
                              col_names = c("edad", "sexo"),
                              percentage = FALSE)
plot_age_sex(data_grouped,
             col_names = c("edad", "sexo"),
             percentage = FALSE)
}
