% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dept_map}
\alias{plot_dept_map}
\title{Plot Department Map by Selected Disease}
\usage{
plot_dept_map(data_map_depto, var_lj = "id")
}
\arguments{
\item{data_map_depto}{Data for department}
}
\value{
The map for department
}
\description{
Función que genera el mapa por departamento de la información de la enfermedad seleccionada
Function that generates the map by information department of the selected disease
}
\examples{
plot_dept_map(data_map_disease_deptos, var_lj = "id")
}
