% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dptos}
\alias{plot_dptos}
\title{Generar gráfico de distribución de casos por departamentos}
\usage{
plot_dptos(data_agrupada, nomb_col = NULL, fuente_data = NULL)
}
\arguments{
\item{data_agrupada}{Un `data.frame` que contiene los datos de la
enfermedad o evento agrupados}

\item{nomb_col}{Un `character` (cadena de carácteres) con el nombre de
la columna de los datos agrupados de la enfermedad o evento que contiene
los departamentos; su valor por defecto es `NULL`}

\item{fuente_data}{Un `character` (cadena de caracteres) que contiene la
leyenda o fuente de información de los datos; su valor por defecto es `NULL`}
}
\value{
Un `plot` o gráfico de distribución de casos por departamentos
}
\description{
Función que genera el gráfico de distribución de casos por departamentos
}
\examples{
data(dengue2020)
data_limpia <- limpiar_data_sivigila(dengue2020)
data_limpia <- estandarizar_geo_cods(data_limpia)
data_agrupada <- agrupar_mpio(data_event = data_limpia,
                              dpto = "Antioquia")
plot_dptos(data_agrupada,
           nomb_col = "departamento_ocurrencia")
}
