% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_edad_sex}
\alias{plot_edad_sex}
\title{Generar gráfico de distribución de casos por edad y sexo}
\usage{
plot_edad_sex(
  data_agrupada,
  col_nombres = c("edad", "sexo"),
  porcentaje = FALSE
)
}
\arguments{
\item{data_agrupada}{Un data frame que contiene los datos de la
enfermedad o evento agrupados}

\item{col_nombres}{Un array (arreglo) de character (cadena de caracteres) con
los nombres de columna de los datos agrupados de la enfermedad o evento que
contiene las edades y las semanas epidemiológicas; su valor por defecto
es c("edad", "sexo")}

\item{porcentaje}{Un boolean (TRUE/FALSE) que indica si los datos tienen
porcentajes; su valor por defecto es FALSE}
}
\value{
Un plot o gráfico de distribución de casos por edad y sexo
}
\description{
Función que genera el gráfico de distribución de casos por edad y sexo
}
\examples{
data_event <- import_data_event(2020, "DENGUE")
data_event <- limpiar_data_sivigila(data_event)
data_agrupada <- agrupar_edad_sex(data_event,
                                  col_nombres = c("edad", "sexo"),
                                  porcentaje = FALSE)
plot_edad_sex(data_agrupada = data_agrupada,
              col_nombres = c("edad", "sexo"),
              porcentaje = FALSE)
}
