% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_fecha_inisintomas}
\alias{plot_fecha_inisintomas}
\title{Generar gráfico de distribución de casos por fecha de inicio de síntomas}
\usage{
plot_fecha_inisintomas(
  data_agrupada,
  col_nombre = "ini_sin",
  uni_marca = "month"
)
}
\arguments{
\item{data_agrupada}{Los datos de la enfermedad agrupados}

\item{col_nombre}{Nombre de la columna en los datos de la
enfermedad agrupados que contiene
las fechas de inicio de síntomas}

\item{uni_marca}{Unidad de marcas (día, mes y año)}
}
\value{
Un gráfico de distribución de casos por fecha de inicio
de síntomas
}
\description{
Función que genera el gráfico de distribución de casos
por fecha de inicio de síntomas
}
\examples{
data_event <- import_data_event(2020, "DENGUE")
data_event <- limpiar_encabezado(data_event)
data_agrupada <- agrupar_fecha_inisintomas(
                                     data_event,
                                     col_nombre = "ini_sin",
                                     tipo = "month")
plot_fecha_inisintomas(data_agrupada,
                       col_nombre = "ini_sin",
                       uni_marca = "month")
}
