% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gender}
\alias{plot_gender}
\title{Plot cases distribution by gender}
\usage{
plot_gender(disease_data, col_name = "sexo", percentage = TRUE)
}
\arguments{
\item{disease_data}{The disease data}

\item{col_name}{Column name in the disease data that contains the gender}
}
\value{
A plot of cases distribution by gender
}
\description{
Function that generates the plot of cases distribution by gender
}
\examples{
disease_data <- import_data_disease_by_year(2020, "DENGUE")
disease_data <- group_by_gender(disease_data, col_name = "sexo", percentage = TRUE)
plot_gender(disease_data, col_name = "sexo", percentage = TRUE)
}
