% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gender_and_week}
\alias{plot_gender_and_week}
\title{plot_gender_and_week}
\usage{
plot_gender_and_week(
  disease_data,
  col_names = c("sexo", "semana"),
  percentage = F
)
}
\arguments{
\item{disease_data}{Disease data}

\item{col_names}{Data set column names}
}
\value{
A plot of cases distribution by gender and epiweek
}
\description{
Función que genera la gráficas de distribucion de casos por genero y semana epidemiologica
Function that generates the plot of cases distribution by gender and epiweek
}
\examples{
disease_data <-  import_data_disease_by_year(2020, "DENGUE")
plot_gender_and_week(disease_data, col_names = c("sexo", "semana"), percentage = F)
}
