% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gender_epiweek}
\alias{plot_gender_epiweek}
\title{Plot cases distribution by gender and epidemiological week}
\usage{
plot_gender_epiweek(
  disease_data,
  col_names = c("sexo", "semana"),
  percentage = FALSE
)
}
\arguments{
\item{disease_data}{The disease data}

\item{col_names}{Column names in the disease data that contains the gender and the epidemiological weeks}
}
\value{
A plot of cases distribution by gender and epidemiological week
}
\description{
Function that generates the plot of cases distribution by gender and epidemiological week
}
\examples{
disease_data <- import_data_disease_by_year(2020, "DENGUE")
disease_data <- group_by_gender_and_week(disease_data, col_names = c("sexo", "semana"), percentage = TRUE)
plot_gender_epiweek(disease_data, col_names = c("sexo", "semana"), percentage = FALSE)
}
