% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_map}
\alias{plot_map}
\title{Generar mapa por departamento}
\usage{
plot_map(
  data_agrupada,
  col_nombre_lj = "id",
  fuente_data = NULL,
  dpto = NULL,
  munpio = NULL
)
}
\arguments{
\item{data_agrupada}{Los datos de la enfermedad agrupados por departamento
y número de casos}

\item{col_nombre_lj}{Nombre de la columna para unir con el archivo
de forma (shape file)}

\item{fuente_data}{Leyenda o fuente de información de los datos
de la enfermedad}

\item{dpto}{Nombre del departamento}

\item{munpio}{Nombre del municipio}
}
\value{
El mapa por departamentos o municipios con el número de casos de una
enfermedad específica
}
\description{
Función que genera el mapa por departamentos o municipios con el número de
casos de una enfermedad específica
}
\examples{
data_event <- import_data_event(2019, "DENGUE")
data_event <- limpiar_encabezado(data_event)
data_espacial_dpto <- estandarizar_geo_cods(data_event)
data_espacial_dpto <- geo_filtro(data_event = data_event,
                                 nombre_dpto = "Antioquia")
data_espacial_dpto <- agrupar_mun(data_event, dept_nombre = "Antioquia")
plot_map(data_espacial_dpto,
   col_nombre_lj = "id",
   fuente_data = "Fuente: SIVIGILA, Instituto Nacional de Salud, Colombia",
   dpto = "Antioquia",
   munpio = "Envigado")
}
