% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_map_dpto}
\alias{plot_map_dpto}
\title{Generar mapa por departamento}
\usage{
plot_map_dpto(data_agrupada, col_nombre_lj = "id", fuente_data = NULL)
}
\arguments{
\item{data_agrupada}{Los datos de la enfermedad agrupados por departamento
y número de casos}

\item{col_nombre_lj}{Nombre de la columna para unir con el archivo
de forma (shape file)}

\item{fuente_data}{Leyenda o fuente de información de los datos de la
enfermedad}
}
\value{
El mapa por departamento con el número de casos de una enfermedad
específica
}
\description{
Función que genera el mapa por departamento con el número de casos de una
enfermedad específica
}
\examples{
data_event <- import_data_event(2019, "DENGUE")
data_event <- limpiar_encabezado(data_event)
data_event <- estandarizar_geo_cods(data_event)
data_dpto <- agrupar_dpto(data_event)
plot_map_dpto(data_dpto,
   col_nombre_lj = "id",
   fuente_data = "Fuente: SIVIGILA, Instituto Nacional de Salud, Colombia")
}
