% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_muns}
\alias{plot_muns}
\title{Generar gráfico de distribución de casos por municipios}
\usage{
plot_muns(data_agrupada, col_nombre = "nombre", porcentaje = FALSE)
}
\arguments{
\item{data_agrupada}{Los datos de la enfermedad agrupados}

\item{col_nombre}{Nombres de columna en los datos de la
enfermedad o evento agrupados que contienen los municipios}

\item{porcentaje}{Indica si los datos tienen porcentajes}
}
\value{
Un gráfico de distribución de casos por municipios
}
\description{
Función que genera el gráfico de distribución de casos por municipios
}
\examples{
data_event <- import_data_event(2020, "DENGUE")
data_event <- limpiar_encabezado(data_event)
data_event <- estandarizar_geo_cods(data_event)
data_agrupada <- agrupar_mun(data_event,
                             dept_nombre = "Antioquia")
plot_muns(data_agrupada,
          col_nombre = "nombre",
          porcentaje = FALSE)
}
