% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_per_etn}
\alias{plot_per_etn}
\title{Generar gráfico de distribución de casos por la pertenencia etnica}
\usage{
plot_per_etn(data_agrupada, nomb_cols = "per_etn", fuente_data = NULL)
}
\arguments{
\item{data_agrupada}{Un `data.frame` que contiene los datos de la
enfermedad o evento agrupados por la la pertenencia etnica}

\item{nomb_cols}{Un `character` (cadena de carácteres) con el nombre de
la columna de los datos agrupados de la enfermedad o evento por la
la pertenencia etnica; su valor por defecto es `"per_etn"`}

\item{fuente_data}{Un `character` (cadena de caracteres) que contiene la
leyenda o fuente de información de los datos; su valor por defecto es `NULL`}
}
\value{
Un `plot` o gráfico de distribución de casos por la clasificación
inicial y los años seleccionados
}
\description{
Función que genera el gráfico por la pertenencia etnica
}
\examples{
data(dengue2020)
data_limpia <- limpiar_data_sivigila(dengue2020)
data_agrupada <- agrupar_per_etn(data_event = data_limpia)
plot_per_etn(data_agrupada,
             nomb_cols = "per_etn")
}
