% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pob_especial}
\alias{plot_pob_especial}
\title{Generar gráfico de distribución de casos por población especial}
\usage{
plot_pob_especial(data_agrupada, col_nombre = "poblacion", porcentaje = FALSE)
}
\arguments{
\item{data_agrupada}{Los datos de la enfermedad agrupados}

\item{col_nombre}{Nombres de columna en los datos de la
enfermedad o evento agrupados que contienen las edades y el sexo}

\item{porcentaje}{Indica si los datos tienen porcentajes}
}
\value{
Un gráfico de distribución de casos por población especial
}
\description{
Función que genera el gráfico de distribución de casos por población especial
}
\examples{
data_event <- import_data_event(2020, "DENGUE")
data_event <- limpiar_encabezado(data_event)
data_agrupada <- agrupar_pob_especial(data_event,
                         col_nombre = "poblacion",
                         porcentaje = TRUE)
plot_pob_especial(data_agrupada,
                  col_nombre = "poblacion",
                  porcentaje = FALSE)
}
