% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_semanaepi}
\alias{plot_semanaepi}
\title{Plot de las semanas epidemiológicas}
\usage{
plot_semanaepi(
  data,
  col_semana,
  col_casos,
  year,
  tipo = "week",
  etiqueta_x = "Semana epidemiologica",
  etiqueta_y = "Numero de casos por semana"
)
}
\arguments{
\item{data}{Los datos de la enfermedad}

\item{col_semana}{Nombre de la columna en los datos de la
enfermedad o evento que contiene las semanas epidemiológicas}

\item{col_casos}{Nombre de la columna en los datos de la
enfermedad o evento que contiene el número de casos}

\item{year}{Año de los datos de la enfermedad}

\item{tipo}{Unidad de tiempo para el eje x (semana y fecha)}

\item{etiqueta_x}{Etiqueta del eje x}

\item{etiqueta_y}{Etiqueta del eje y}
}
\value{
El gráfico de las semanas epidemiológicas
}
\description{
Función que genera el gráfico de las semanas epidemiológicas
}
\examples{
data_resumida_sivigila <- import_data_resumen_sivigila()
data_resumida_sivigila <- limpiar_encabezado(data_resumida_sivigila)
data_filtrada <- filtrar_event("MALAR", data_resumida_sivigila)
plot_semanaepi(data_filtrada,
               col_semana = "semana",
               col_casos = "casos",
               year = 2019,
               tipo = "date",
               etiqueta_x = "Fecha de semana epidemiologica")
}
