% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_sex}
\alias{plot_sex}
\title{Generar gráfico de distribución de casos por sexo}
\usage{
plot_sex(data_agrupada, col_nombre = "sexo", porcentaje = TRUE)
}
\arguments{
\item{data_agrupada}{Los datos de la enfermedad agrupados}

\item{col_nombre}{Nombre de la columna en los datos de la enfermedad
agrupados que contiene
el sexo}

\item{porcentaje}{Indica si los datos tienen porcentajes}
}
\value{
Un gráfico de distribución de casos por sexo
}
\description{
Función que genera el gráfico de distribución de casos por sexo
}
\examples{
data_event <- import_data_event(2020, "DENGUE")
data_event <- limpiar_encabezado(data_event)
data_agrupada <- agrupar_sex(data_event,
                             col_nombre = "sexo",
                             porcentaje = TRUE)
plot_sex(data_agrupada, col_nombre = "sexo", porcentaje = TRUE)
}
