% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R, R/plotting.R
\name{plot_sex_epiweek}
\alias{plot_sex_epiweek}
\title{Plot cases distribution by sex and epidemiological week}
\usage{
plot_sex_epiweek(
  data_grouped,
  col_names = c("sexo", "semana"),
  percentage = FALSE
)

plot_sex_epiweek(
  data_grouped,
  col_names = c("sexo", "semana"),
  percentage = FALSE
)
}
\arguments{
\item{data_grouped}{The disease data grouped}

\item{col_names}{Column names in the disease data grouped that contains
the sex and the epidemiological weeks}

\item{percentage}{Indicates if the data has percentages}
}
\value{
A plot of cases distribution by sex and epidemiological week

A plot of cases distribution by sex and epidemiological week
}
\description{
Function that generates the plot of cases distribution by sex and
epidemiological week

Function that generates the plot of cases distribution by sex and
epidemiological week
}
\examples{
disease_data <- import_linelist_disease_year(2020, "DENGUE")
disease_data <- clean_header(disease_data)
data_grouped <- group_sex_epiweek(disease_data,
                                   col_names = c("sexo", "semana"),
                                   percentage = TRUE)
plot_sex_epiweek(data_grouped,
                 col_names = c("sexo", "semana"),
                 percentage = FALSE)
disease_data <- import_linelist_disease_year(2020, "DENGUE")
disease_data <- clean_header(disease_data)
data_grouped <- group_sex_epiweek(disease_data,
                                   col_names = c("sexo", "semana"),
                                   percentage = TRUE)
plot_sex_epiweek(data_grouped,
                 col_names = c("sexo", "semana"),
                 percentage = FALSE)
}
