% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_special_population}
\alias{plot_special_population}
\title{Plot cases distribution by special population}
\usage{
plot_special_population(
  disease_data,
  col_name = "poblacion",
  percentage = FALSE
)
}
\arguments{
\item{disease_data}{The disease data}

\item{col_name}{Column names in the disease data that contains the ages and gender}
}
\value{
A plot of cases distribution by special population
}
\description{
Function that generates the plot of cases distribution by special population
}
\examples{
disease_data <- import_data_disease_by_year(2020, "DENGUE")
disease_data <- group_by_age_and_gender(disease_data, col_names = c("edad", "sexo"), percentage = TRUE)
plot_special_population(disease_data, col_name = "poblacion", percentage = FALSE)
}
