% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_tabla_tipos_event}
\alias{plot_tabla_tipos_event}
\title{Generar tabla con la distribución de casos por tipo de
enfermedad o evento}
\usage{
plot_tabla_tipos_event(data_agrupada, nomb_col = "nombre_evento")
}
\arguments{
\item{data_agrupada}{Un `data.frame` que contiene los datos de la
enfermedad o evento agrupados}

\item{nomb_col}{Un `character` (cadena de carácteres) con el nombre de
la columna de los datos agrupados de la enfermedad o evento por
tipo; su valor por defecto es `"nombre_evento"`}
}
\value{
Una `kable` (tabla gráfica) con la distribución de casos
por tipo de enfermedad o evento
}
\description{
Función que genera el la tabla con la distribución de casos por
tipo de enfermedad o evento
}
