% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_variable}
\alias{plot_variable}
\title{Generar gráfico por variable(s) o columna(s)}
\usage{
plot_variable(
  data,
  var_x,
  var_y,
  var_por = NULL,
  var_fill = NULL,
  agr_per = TRUE,
  etiqueta_x,
  etiqueta_y,
  nombre_escala = NULL,
  etiquetas_escala = NULL,
  tit_diagram = NULL,
  pos_leyenda,
  ancho_barra = 1,
  tam_text = 3,
  most_val = TRUE,
  fuente_data = NULL
)
}
\arguments{
\item{data}{Los datos a graficar}

\item{var_x}{El nombre de la variable para el eje x}

\item{var_y}{El nombre de la variable para el eje y}

\item{var_por}{El nombre de la variable si los datos tienen porcentajes}

\item{var_fill}{La variable de llenado}

\item{agr_per}{Indica si los datos tienen porcentajes}

\item{etiqueta_x}{La etiqueta para el eje x}

\item{etiqueta_y}{La etiqueta para el eje y}

\item{nombre_escala}{El nombre de la escala}

\item{etiquetas_escala}{Las etiquetas de la escala}

\item{tit_diagram}{El título del diagrama}

\item{pos_leyenda}{La posición de la leyenda}

\item{ancho_barra}{El ancho de las barra}

\item{tam_text}{El tamaño del texto}

\item{most_val}{Indica si las barras deben mostrar los valores}

\item{fuente_data}{Leyenda o fuente de información de los datos}
}
\value{
El gráfico por variable(s) o columna(s)
}
\description{
Función que genera un gráfico por cualquier tipo de variable o columna de un
dataframe
}
\examples{
data_event <- import_data_event(2019, "DENGUE")
data_event <- limpiar_encabezado(data_event)
cases_sex <- agrupar_sex(data_event,
                         porcentaje = TRUE)
plot_variable(cases_sex,
  var_x = "sexo",
  var_y = "casos",
  var_fill = "sexo",
  var_por = "porcentaje",
  etiqueta_x = "Sexo",
  etiqueta_y = "Numero de casos",
  nombre_escala = "Sexo",
  etiquetas_escala = c("Femenino", "Masculino"),
  pos_leyenda = "right",
  ancho_barra = 0.5,
  tam_text = 3,
  most_val = TRUE
  )
}
