% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{remove_cols}
\alias{remove_cols}
\title{Eliminar columnas erróneas de los datos del evento o enfermedad}
\usage{
remove_cols(data_event)
}
\arguments{
\item{data_event}{Un `data.frame` que contiene los datos de una enfermedad
o evento.}
}
\value{
Un `data.frame` con las columnas correctas de los datos del evento o
enfermedad.
}
\description{
Función que elimina las columnas erróneas de los datos
del evento o enfermedad al importarlos desde el API del SIVIGILA.
}
\keyword{internal}
