% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{remove_dates_init}
\alias{remove_dates_init}
\title{Remove Dates Init}
\usage{
remove_dates_init(disease_data, col_init = "INI_SI", col_cmp = "FEC_HOS")
}
\arguments{
\item{disease_data}{Disease data}

\item{col_init}{Column name of initial value}

\item{col_cmp}{Column name of comparison value}
}
\value{
The clean data
}
\description{
Función que elimina los valores que son superiores al valor de comparación
Function that removes values that are greater than the comparison value
}
\examples{
remove_dates_init(disease_data, col_init = "INI_SI", col_cmp = "FEC_HOS")
}
