% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{remove_error_dates}
\alias{remove_error_dates}
\title{Remove error dates}
\usage{
remove_error_dates(disease_data, col_init = "ini_sin", col_cmp = "fec_hos")
}
\arguments{
\item{disease_data}{The disease data}

\item{col_init}{Column name of the date}

\item{col_cmp}{Column name of the comparison date}
}
\value{
The data without the erroneous dates
}
\description{
Function that removes dates greater than the comparison value
}
\examples{
disease_data <- import_linelist_disease_year(2019, "DENGUE")
remove_error_dates(disease_data, col_init = "ini_sin", col_cmp = "fec_hos")
}
