% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{remove_error_fecha}
\alias{remove_error_fecha}
\title{Eliminar fechas erróneas}
\usage{
remove_error_fecha(data_event, col_ini = "ini_sin", col_comp = "fec_hos")
}
\arguments{
\item{data_event}{Los datos de una enfermedad o evento}

\item{col_ini}{Nombre de la columna de la fecha}

\item{col_comp}{Nombre de la columna de la fecha de comparación}
}
\value{
Los datos sin las fechas erróneas
}
\description{
Función que elimina fechas mayores que el valor de comparación
}
\examples{
data_event <- import_data_event(2019, "DENGUE")
data_event <- limpiar_encabezado(data_event)
remove_error_fecha(data_event, col_ini = "ini_sin", col_comp = "fec_hos")
}
