% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{remove_nin_values}
\alias{remove_nin_values}
\title{Remove NIN (NA, Infinitive, NaN) Values}
\usage{
remove_nin_values(disease_data, name_col)
}
\arguments{
\item{disease_data}{Disease data}

\item{name_col}{Column name}
}
\value{
The clean data without NA, Infinitive or NaN values of the column
}
\description{
Función que elimina los valores (NA, Infinitive, NaN) de una columna\n
Function that eliminates the values (NA, Infinitive, NaN) of a column
}
\examples{
remove_nin_values(disease_data, name_col = "EDAD")
}
