% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{remove_val_nin}
\alias{remove_val_nin}
\title{Eliminar valores NIN (NA, Infinito, NaN)}
\usage{
remove_val_nin(data_event, nomb_col)
}
\arguments{
\item{data_event}{Un `data.frame` que contiene los datos de una
enfermedad o evento.}

\item{nomb_col}{Un `character` (cadena de caracteres) que contiene el
nombre de la columna a evaluar en los datos de una enfermedad o evento.}
}
\value{
Un `data.frame` con los datos limpios sin valores NA,
Infinito o NaN.
}
\description{
Función que elimina filas si los valores de la columna
seleccionada incluyen NA, Infinito o NaN.
}
\keyword{internal}
