% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{select_disease}
\alias{select_disease}
\title{Select Disease}
\usage{
select_disease(name_disease, data)
}
\arguments{
\item{name_disease}{Name of disease}

\item{data}{data set}
}
\value{
The filtered data with the disease selected information
}
\description{
Función que seleccione una enfermedad por ejemplo Tuberculosis
Function that selects a disease for example Tuberculosis
}
\examples{
select_disease("MALARIA", data_sivigila)
}
