% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_code_mun}
\alias{set_code_mun}
\title{Set geographic codes of the disease data}
\usage{
set_code_mun(code_dept, code_mun)
}
\arguments{
\item{code_dept}{The department code}

\item{code_mun}{The municipality code}
}
\value{
The geographic codes of the disease data
}
\description{
Function that sets the geographic codes of the disease data
}
\examples{
set_code_mun(code_dept = 01, code_mun = "001")
}
