% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sivirep-package.R
\docType{package}
\name{sivirep-package}
\alias{sivirep}
\alias{sivirep-package}
\title{sivirep: Data Wrangling and Automated Reports from 'SIVIGILA' source}
\description{
Provides functions for data wrangling and automated reports from 'SIVIGILA' source.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://epiverse-trace.github.io/sivirep/}
  \item \url{https://github.com/epiverse-trace/sivirep}
  \item Report bugs at \url{https://github.com/epiverse-trace/sivirep/issues}
}

}
\author{
\strong{Maintainer}: Zulma M. Cucunubá \email{zulma.cucunuba@javeriana.edu.co} (\href{https://orcid.org/0000-0002-8165-3198}{ORCID}) [contributor]

Authors:
\itemize{
  \item Geraldine Gómez-Millán \email{geralidine.gomez@javeriana.edu.co} (\href{https://orcid.org/0009-0007-8701-0568}{ORCID}) [contributor]
  \item Jennifer A. Mendez-Romero \email{jenniferk2@hotmail.com} (\href{https://orcid.org/0009-0001-6138-0225}{ORCID}) [contributor]
  \item Claudia Huguett-Aragón \email{chuguett@hotmail.com} (\href{https://orcid.org/0000-0002-9814-2386}{ORCID}) [contributor]
}

Other contributors:
\itemize{
  \item Hugo Gruson \email{hugo.gruson+R@normalesup.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [contributor]
  \item Juanita Romero-Garcés \email{juanita.romerog@javeriana.edu.co} (\href{https://orcid.org/0000-0001-9796-6626}{ORCID}) [contributor]
  \item Jaime Pavlich-Mariscal \email{jpavlich@javeriana.edu.co} (\href{https://orcid.org/0000-0002-3892-6680}{ORCID}) [contributor]
  \item Laura Gómez Bermeo \email{gomezblaura@javeriana.edu.co} (\href{https://orcid.org/0000-0003-4028-2893}{ORCID}) [contributor]
  \item Andrés Moreno \email{ad.morenob@javeriana.edu.co} (\href{https://orcid.org/0000-0001-9266-731X}{ORCID}) [contributor]
  \item Miguel Gámez [contributor]
  \item Johan Calderón [contributor]
  \item Lady Flórez-Tapiero [contributor]
  \item Verónica Tangarife-Arredondo [contributor]
  \item Gerard Alarcon [contributor]
  \item International Development Research Center (IDRC) [funder]
  \item Pontificia Universidad Javeriana [copyright holder]
}

}
\keyword{internal}
