% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{standardize_geo_codes}
\alias{standardize_geo_codes}
\title{Standardize geographic codes of the disease data}
\usage{
standardize_geo_codes(disease_data)
}
\arguments{
\item{disease_data}{The disease data}
}
\value{
The geographic codes of the disease data
}
\description{
Function that standardizes the geographic codes of the disease data
}
\examples{
disease_data <- import_linelist_disease_year(2019, "DENGUE")
disease_data <- clean_header(disease_data)
standardize_geo_codes(disease_data)
}
