% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_sivirep.R
\name{tema_sivirep}
\alias{tema_sivirep}
\title{Tema sivirep}
\usage{
tema_sivirep()
}
\value{
Un objeto tema de \pkg{ggplot2}.
}
\description{
Tema \pkg{ggplot2} personalizado para los reportes
de \pkg{sivirep}.
}
\keyword{internal}
