% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{agrupar_area_geo}
\alias{agrupar_area_geo}
\title{Agrupar por área geográfica}
\usage{
agrupar_area_geo(
  data_event,
  dpto = NULL,
  nomb_col = "area",
  porcentaje = FALSE
)
}
\arguments{
\item{data_event}{Un `data.frame` que contiene los datos de la
enfermedad o evento}

\item{dpto}{Un `character` (cadena de caracteres) que contiene
el nombre del departamento; su valor por defecto es `NULL`; si se ingresa
un valor en este parámetro se procederá agrupar los datos
por los municipios del departamento y sus áreas geográficas; si no se
ingresa un valor en este parámetro validará si los datos ya están
filtrados por algún departamento, si no lo están generará la agrupación
por departamento}

\item{nomb_col}{Un `character` (cadena de caracteres) con el nombre de
la columna en los datos de la enfermedad o evento que contiene los códigos
de municipios; su valor por defecto es `"cod_mun_o"`}

\item{porcentaje}{Un `boolean` (TRUE o FALSE) que indica si es necesario
agregar un porcentaje de casos como una columna; su valor por
defecto es `FALSE`}
}
\value{
Un `data.frame` con los datos de la enfermedad o evento agrupados
por códigos de municipios y número de casos
}
\description{
Función que agrupa los datos de una enfermedad o evento por área
geográfica a nivel departamental o municipal
}
\examples{
data(dengue2020)
data_limpia <- limpiar_data_sivigila(data_event = dengue2020)
agrupar_area_geo(data_event = data_limpia,
                 dpto = "Antioquia",
                 nomb_col = "area",
                 porcentaje = FALSE)
}
