% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{agrupar_cols_casos}
\alias{agrupar_cols_casos}
\title{Agrupar por columnas y casos}
\usage{
agrupar_cols_casos(data_event, cols_nombres, agr_porcentaje = FALSE)

agrupar_cols_casos(data_event, cols_nombres, agr_porcentaje = FALSE)
}
\arguments{
\item{data_event}{Los datos de una enfermedad o evento}

\item{cols_nombres}{El nombre de la(s) columna(s)}

\item{agr_porcentaje}{Indica si es necesario agregar un porcentaje de
casos como una columna}
}
\value{
Los datos de una enfermedad o evento agrupados
por un nombre de columna(s) específico y número de casos

Los datos de una enfermedad o evento agrupados
por un nombre de columna(s) específico y número de casos
}
\description{
Función que agrupa los datos de una enfermedad o evento
por un nombre de columna(s) específico y número de casos

Función que agrupa los datos de una enfermedad o evento
por un nombre de columna(s) específico y número de casos
}
\examples{
data_event <- import_data_event(2019, "DENGUE")
data_event <- limpiar_encabezado(data_event)
agrupar_cols_casos(data_event,
                    cols_nombres = "sexo",
                    agr_porcentaje = TRUE)
agrupar_cols_casos(data_event, cols_nombres = c("sexo", "semana"))
data_event <- import_data_event(2019, "DENGUE")
data_event <- limpiar_encabezado(data_event)
agrupar_cols_casos(data_event, cols_nombres = "sexo", agr_porcentaje = TRUE)
agrupar_cols_casos(data_event, cols_nombres = c("sexo", "semana"))
}
