% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{agrupar_dpto}
\alias{agrupar_dpto}
\title{Agrupar por departamento y casos}
\usage{
agrupar_dpto(data_event, col_nombre = "cod_dpto_o", porcentaje = FALSE)
}
\arguments{
\item{data_event}{Un data frame que contiene los datos de la
enfermedad o evento}

\item{col_nombre}{Un character (cadena de caracteres) con el nombre
de la columna en los datos de la enfermedad o evento que contiene los
códigos de departamento; su valor por defecto es cod_dpto_o}

\item{porcentaje}{Un boolean (TRUE o FALSE) que indica si
es necesario agregar un porcentaje de casos como una columna; su valor
por defecto es FALSE}
}
\value{
Un data frame con los datos de la enfermedad o evento agrupados
por códigos de departamento y número de casos
}
\description{
Función que agrupa los datos por códigos de departamento y
número de casos
}
\examples{
data_event <- import_data_event(2019, "DENGUE")
data_event <- limpiar_encabezado(data_event)
agrupar_dpto(data_event = data_event,
             col_nombre = "cod_dpto_o",
             porcentaje = FALSE)
}
