% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{agrupar_edad}
\alias{agrupar_edad}
\title{Agrupar por edad y casos}
\usage{
agrupar_edad(
  data_event,
  col_nombre = "edad",
  porcentaje = FALSE,
  interval_edad = 10
)
}
\arguments{
\item{data_event}{Los datos de enfermedades}

\item{col_nombre}{Nombre de la columna en los datos de enfermedades que
contiene las edades}

\item{porcentaje}{Indica si es necesario agregar un porcentaje de
casos como una columna}

\item{interval_edad}{El intervalo del rango de edades}
}
\value{
Los datos de enfermedades agrupados por edad y número de casos
}
\description{
Función que agrupa los datos de enfermedades por edad y número de casos
}
\examples{
data_event <- import_data_event(2019, "DENGUE")
data_event <- limpiar_encabezado(data_event)
agrupar_edad(data_event, col_nombre = "edad", porcentaje = FALSE)
}
