% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{agrupar_edad}
\alias{agrupar_edad}
\title{Agrupar por edad y casos}
\usage{
agrupar_edad(
  data_event,
  col_nombre = "edad",
  porcentaje = FALSE,
  interval_edad = 10
)
}
\arguments{
\item{data_event}{Un data frame que contiene los datos de la enfermedad
o evento}

\item{col_nombre}{Un character (cadena de caracteres) con el nombre
de la columna de los datos de la enfermedad o evento que contiene las edades;
su valor por defecto es edad}

\item{porcentaje}{Un boolean (TRUE o FALSE) que indica si
es necesario agregar un porcentaje de casos como una columna; su valor por
defecto es `FALSE`}

\item{interval_edad}{Un numeric (numerico) que contiene el intervalo del
rango de edades; su valor por defecto es 10}
}
\value{
Un data frame con los datos de la enfermedad o evento agrupados
por edad y número de casos
}
\description{
Función que agrupa los datos de una enfermedad o evento por edad
y número de casos
}
\examples{
data_event <- import_data_event(2019, "DENGUE")
data_event <- limpiar_encabezado(data_event)
agrupar_edad(data_event = data_event,
             col_nombre = "edad",
             porcentaje = FALSE)
}
