% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{agrupar_edad_sex}
\alias{agrupar_edad_sex}
\title{Agrupar por edades, sexo y casos}
\usage{
agrupar_edad_sex(
  data_event,
  col_nombres = c("edad", "sexo"),
  porcentaje = TRUE,
  interval_edad = 10
)
}
\arguments{
\item{data_event}{Los datos de enfermedades}

\item{col_nombres}{Nombres de las columnas en los datos de
enfermedades que contienen las edades y el sexo}

\item{porcentaje}{Indica si es necesario agregar un porcentaje de
casos como una columna}

\item{interval_edad}{El intervalo del rango de edades}
}
\value{
Los datos de enfermedades agrupados por edades, sexo y
número de casos
}
\description{
Función que agrupa los datos de enfermedades por edades,
sexo y número de casos
}
\examples{
data_event <- import_data_event(2019, "DENGUE")
data_event <- limpiar_encabezado(data_event)
agrupar_edad_sex(data_event,
                 col_nombres = c("edad", "sexo"),
                 porcentaje = TRUE)
}
