% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{agrupar_fecha_notifica}
\alias{agrupar_fecha_notifica}
\title{Agrupar por fecha de notificación y casos}
\usage{
agrupar_fecha_notifica(data_event, col_nombre = "fec_not", tipo = "month")
}
\arguments{
\item{data_event}{Los datos de enfermedades}

\item{col_nombre}{Nombre de la columna en los datos de enfermedades
que contiene las fechas de notificación}

\item{tipo}{Unidad de tiempo (día, mes y año)}
}
\value{
Los datos de enfermedades agrupados por fecha de
notificación y número de casos
}
\description{
Función que agrupa los datos de enfermedades por fecha de
notificación y número de casos
}
\examples{
data_event <- import_data_event(2019, "DENGUE")
data_event <- limpiar_encabezado(data_event)
agrupar_fecha_notifica(data_event, col_nombre = "fec_not", tipo = "month")
}
