% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{agrupar_pob_especial}
\alias{agrupar_pob_especial}
\title{Agrupar por población especial y casos}
\usage{
agrupar_pob_especial(data_event, col_nombre = "poblacion", porcentaje = TRUE)
}
\arguments{
\item{data_event}{Los datos de la enfermedad}

\item{col_nombre}{Nombre de la columna en los datos de la
enfermedad que contiene las poblaciones especiales}

\item{porcentaje}{Indica si se requiere agregar un
porcentaje de casos como columna}
}
\value{
Los datos de la enfermedad agrupados por poblaciones
especiales y casos
}
\description{
Función que agrupa los datos de la enfermedad por población especial y casos
}
\examples{
data_event <- import_data_event(2019, "DENGUE")
data_event <- limpiar_encabezado(data_event)
agrupar_pob_especial(data_event,
                     col_nombre = "poblacion",
                     porcentaje = TRUE)
}
