% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{agrupar_pob_especial}
\alias{agrupar_pob_especial}
\title{Agrupar por población especial y casos}
\usage{
agrupar_pob_especial(data_event, col_nombre = "poblacion", porcentaje = TRUE)
}
\arguments{
\item{data_event}{Un data frame que contiene los datos de la enfermedad
o evento}

\item{col_nombre}{Un character (cadena de caracteres) con el nombre de la
columna de los datos de la enfermedad o evento que contiene las poblaciones
especiales; su valor por defecto es poblacion}

\item{porcentaje}{Un boolean (TRUE o FALSE) que indica si
es necesario agregar un porcentaje de casos como una columna; su valor
por defecto es `TRUE`}
}
\value{
Un data frame con los datos de la enfermedad o evento agrupados
por poblaciones especiales y casos
}
\description{
Función que agrupa los datos de la enfermedad o evento por población
especial y casos
}
\examples{
data_event <- import_data_event(2019, "DENGUE")
data_event <- limpiar_encabezado(data_event)
agrupar_pob_especial(data_event = data_event,
                     col_nombre = "poblacion",
                     porcentaje = TRUE)
}
