% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{agrupar_rango_edad_casos}
\alias{agrupar_rango_edad_casos}
\title{Agrupar por rango de edad y casos}
\usage{
agrupar_rango_edad_casos(
  data_event,
  col_nombre,
  var_a = NULL,
  min_val,
  max_val,
  paso
)
}
\arguments{
\item{data_event}{Los datos de una enfermedad o evento}

\item{col_nombre}{Nombre de la columna en los datos de una enfermedad
o evento que contiene las edades}

\item{var_a}{Nombre adicional de columna en los datos de
una enfermedad o evento para agrupar con la edad}

\item{min_val}{Valor mínimo de las edades}

\item{max_val}{Valor máximo de las edades}

\item{paso}{Paso para generar el rango de edades}
}
\value{
Los datos de una enfermedad o evento agrupados por el rango
de edad y número de casos
}
\description{
Función que agrupa los datos de una enfermedad o evento por rango
de edad y número de casos
}
\examples{
data_event <- import_data_event(2019, "DENGUE")
data_event <- limpiar_encabezado(data_event)
disease_dt_by_age <- agrupar_cols_casos(data_event, c("edad", "semana"),
  agr_porcentaje = TRUE
)
agrupar_rango_edad_casos(disease_dt_by_age, "edad",
  min_val = 0,
  max_val = max(disease_dt_by_age$edad), paso = 10
)
}
