% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{agrupar_rango_edad_casos}
\alias{agrupar_rango_edad_casos}
\title{Agrupar por rango de edad y casos}
\usage{
agrupar_rango_edad_casos(
  data_event,
  col_nombre,
  var_a = NULL,
  min_val,
  max_val,
  paso
)
}
\arguments{
\item{data_event}{Un data frame que contiene los datos de la
enfermedad o evento}

\item{col_nombre}{Un character (cadena de caracteres) con
el nombre de la columna de los datos de la enfermedad o evento
que contiene las edades}

\item{var_a}{Un character (cadena de caracteres) con
el nombre adicional de la columna de los datos de la enfermedad
o evento para agrupar con la edad; su valor por defecto es
NULL}

\item{min_val}{Un numeric (numerico) que contiene el valor mínimo
de las edades}

\item{max_val}{Un numeric (numerico) que contiene el valor máximo
de las edades}

\item{paso}{Un numeric (numerico) que contiene el valor del paso
para generar el rango de edades}
}
\value{
Un data frame con los datos de la enfermedad o evento
agrupados por el rango de edad y número de casos
}
\description{
Función que agrupa los datos de una enfermedad o evento por rango
de edad y número de casos
}
\examples{
data(dengue2020)
data_event <- dengue2020
data_event <- limpiar_data_sivigila(data_event, 2020)
data_edad <- agrupar_cols_casos(data_event = data_event,
                                c("edad", "semana"),
                                agr_porcentaje = TRUE)
agrupar_rango_edad_casos(data_event = data_edad,
                         col_nombre = "edad",
                         min_val = 0,
                         max_val = max(data_edad$edad),
                         paso = 10)
}
