% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{agrupar_sex}
\alias{agrupar_sex}
\title{Agrupar por sexo y casos}
\usage{
agrupar_sex(data_event, nomb_col = "sexo", porcentaje = TRUE)
}
\arguments{
\item{data_event}{Un `data.frame` que contiene los datos de la enfermedad
o evento}

\item{nomb_col}{Un `character` (cadena de caracteres) con el nombre
de la columna de los datos de la enfermedad o evento que contiene el sexo;
su valor por defecto es `"sexo"`}

\item{porcentaje}{Un `boolean` (TRUE o FALSE) que indica si es necesario
agregar un porcentaje de casos como una columna; su valor por defecto es
`TRUE`}
}
\value{
Un data.frame con los datos de la enfermedad o evento
agrupados por sexo y número de casos
}
\description{
Función que agrupa los datos de una enfermedad o evento
por sexo y número de casos
}
\examples{
data(dengue2020)
data_limpia <- limpiar_data_sivigila(data_event = dengue2020)
agrupar_sex(data_event = data_limpia,
            nomb_col = "sexo",
            porcentaje = TRUE)
}
