% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{agrupar_years}
\alias{agrupar_years}
\title{Agrupar por tipo de enfermedad o evento}
\usage{
agrupar_years(data_event, nomb_col = "ano")
}
\arguments{
\item{data_event}{Un `data.frame` que contiene los datos de la
enfermedad o evento}

\item{nomb_col}{Un `character` (cadena de caracteres) con el nombre de
la columna en los datos de la enfermedad o evento que contiene los a침os
del evento; su valor por defecto es `"ano"`}
}
\value{
Un `data.frame` con los datos de la enfermedad o evento agrupados
por sus a침os
}
\description{
Funci칩n que agrupa los casos por los a침os de una enfermedad o evento
}
\examples{
data(dengue2020)
data_limpia <- limpiar_data_sivigila(data_event = dengue2020)
agrupar_years(data_event = data_limpia,
              nomb_col = "ano")
}
