% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{calcular_incidencia}
\alias{calcular_incidencia}
\title{Calcular incidencia}
\usage{
calcular_incidencia(
  data_incidencia,
  data_agrupada,
  year,
  dpto = NULL,
  mpio = NULL
)
}
\arguments{
\item{data_incidencia}{Un `data.frame` que contiene la proyecciones
poblaciones del DANE}

\item{data_agrupada}{Un `data.frame` que contiene los datos de la enfermedad
agrupados por departamento o municipio y número de casos}

\item{dpto}{Un `character` (cadena de caracteres) o `numeric` (numérico)
que contiene el código o nombre del departamento; su valor por
defecto es `NULL`}

\item{mpio}{Un `character` (cadena de caracteres) o `numeric` (numérico)
que contiene el código o nombre del municipio; su valor por defecto es `NULL`}

\item{years}{Un `numeric` (numerico) con el año que se debe tomar de las
proyecciones poblacionales}
}
\value{
Un `numeric` con el calculo de la incidencia para todo Colombia, un
departamento o municipio
}
\description{
Función que calcula la incidencia de una enfermedad o evento para todo
Colombia, departamento o municipio
}
\examples{
\dontrun{
data(dengue2020)
data_limpia <- limpiar_data_sivigila(data_event = dengue2020)
proyecciones <- import_data_inicidencia()
proyecciones_limpias <- limpiar_data_incidencia(data_incidencia =
                                                 proyecciones)
data_agrupada_mpios <- agrupar_mpio(data_limpia, dpto = "Antioquia")
calcular_incidencia(data_incidencia = proyecciones_limpias,
                    data_agrupada = data_agrupada_mpios,
                    dpto = "05",
                    year = 2020)
calcular_incidencia(data_incidencia = proyecciones_limpias,
                    data_agrupada = data_agrupada_mpios,
                    dpto = "05",
                    mpio = "05001"
                    year = 2020)
data_agrupada_dptos <- agrupar_dpto(data_limpia)
calcular_incidencia(data_incidencia = proyecciones_limpias,
                    data_agrupada = data_agrupada_dptos,
                    year = 2020)
}
}
