% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{clean_depto_codes}
\alias{clean_depto_codes}
\title{Clean department codes of the disease data}
\usage{
clean_depto_codes(disease_data, col_data_codes, geo_data, col_geo_codes)
}
\arguments{
\item{disease_data}{The disease data}

\item{col_data_codes}{Column name in the disease data that contains the department codes}

\item{col_geo_codes}{Column name in the geographical data than contains the department codes}

\item{geo_codes}{Geographical data that includes the departments codes}
}
\value{
Clean department codes of the disease data
}
\description{
Function that cleans the department codes of the disease data
}
\examples{
geo_codes <- import_geo_codes()
disease_data <- import_data_disease_by_year(2019, "DENGUE")
clean_depto_codes(disease_data, col_data_codes = "cod_dpto_o", geo_codes, col_geo_codes = "c_digo_departamento")
}
