% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{clean_depto_disease_codes}
\alias{clean_depto_disease_codes}
\title{Clean department codes of the disease data}
\usage{
clean_depto_disease_codes(depto_codes, disease_data, make_group = TRUE)
}
\arguments{
\item{depto_codes}{The department codes}

\item{disease_data}{The disease data}

\item{make_group}{Indicates if it is necessary to group by 
department IDs and case numbers}
}
\value{
Clean department codes of the disease data
}
\description{
Function that cleans the department codes of the disease data
}
\examples{
geo_codes <- import_geo_codes()
depto_codes <- get_depto_codes(geo_codes)
disease_data <- import_linelist_disease_year(2019, "DENGUE")
data_grouped <- group_columns_cases(disease_data, "COD_DPTO_O", wt_percentage = TRUE)
clean_depto_disease_codes(data_grouped, disease_data)
}
