% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{clean_disease_ages}
\alias{clean_disease_ages}
\title{clean_disease_dates}
\usage{
clean_disease_ages(disease_data, col_name = "edad")
}
\arguments{
\item{disease_data}{Disease data}

\item{col_name}{Data set column name}
}
\value{
The data with clean ages
}
\description{
Función que limpia las edades de un conjunto de datos
Function that cleans ages from a dataset
}
\examples{
disease_data <- import_data_disease_by_year(2020, "DENGUE")
clean_disease_ages(disease_data, col_name = "edad")
}
