% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{clean_disease_dates}
\alias{clean_disease_dates}
\title{clean_disease_dates}
\usage{
clean_disease_dates(
  disease_data,
  year,
  date_format = "\%AAAA-\%MM-\%DD",
  col_name = "ini_sin",
  col_cmp = NULL
)
}
\arguments{
\item{disease_data}{Disease data}

\item{date_format}{Date format}

\item{col_name}{Data set column name}

\item{col_cmp}{Data set column compare}
}
\value{
The data with clean dates
}
\description{
Función que limpia las fechas de un conjunto de datos
Function that cleans dates from a dataset
}
\examples{
disease_data <- import_data_disease_by_year(2020, "DENGUE")
clean_disease_dates(disease_data, year, date_format = "\%AAAA-\%MM-\%DD", col_name = "ini_sin", col_cmp = "fec_hos")
}
