% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{clean_disease_dates}
\alias{clean_disease_dates}
\title{Clean dates of the disease data}
\usage{
clean_disease_dates(
  disease_data,
  year,
  date_format = "\%AAAA-\%MM-\%DD",
  col_name = "ini_sin",
  col_cmp = NULL
)
}
\arguments{
\item{disease_data}{The disease data}

\item{year}{Year of the disease data}

\item{date_format}{Date format}

\item{col_name}{Data set column name}

\item{col_cmp}{Data set column compare}
}
\value{
The data with clean dates
}
\description{
Function that cleans dates from the disease data
}
\examples{
disease_data <- import_linelist_disease_year(2020, "DENGUE")
disease_data <- clean_header(disease_data)
clean_disease_dates(disease_data, 
                     year = 2020, 
                     date_format = "\%AAAA-\%MM-\%DD", 
                     col_name = "ini_sin", 
                     col_cmp = "fec_hos")
}
