% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{clean_sivigila_data}
\alias{clean_sivigila_data}
\title{Clean SIVIGILA data}
\usage{
clean_sivigila_data(disease_data, year)
}
\arguments{
\item{disease_data}{The disease data}

\item{year}{Year of the disease data}
}
\value{
Clean data of the disease
}
\description{
Function that cleans the selected disease data of SIVIGILA source
}
\examples{
year <- 2019
disease_data <- import_data_disease_by_year(year, "DENGUE")
clean_sivigila_data(disease_data, year)
}
