% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dengue2020.R
\docType{data}
\name{dengue2020}
\alias{dengue2020}
\title{Datos Dengue 2020 en sivirep}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
dengue2020
}
\description{
Data from
}
\examples{
dengue2020
}
\keyword{datasets}
