% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divipoladata.R
\docType{data}
\name{divipoladata}
\alias{divipoladata}
\title{Códigos e información geografica del DIVIPOLA en sivirep}
\format{
## `divipoladata`
A `<data.frame>` con 1121 filas and 5 columnas:
\describe{
  \item{codigo_departamento}{Codigo de los departamentos}
  \item{codigo_municipio}{Codigo de los municipios}
  \item{nombre_departamento}{Nombre de los departamentos}
  \item{nombre_municipio}{Nombre de los municipios}
  \item{tipo_municipio_isla_area_no_municipalizada}{Tipo}
}
}
\usage{
data(divipoladata)
}
\description{
Datos obtenidos de la API de Datos Abiertos de Colombia.
}
\examples{
data(divipoladata)
}
\keyword{datasets}
