% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{estandarizar_geo_cods}
\alias{estandarizar_geo_cods}
\title{Estandarizar códigos geográficos de los datos de una enfermedad o evento}
\usage{
estandarizar_geo_cods(data_event)
}
\arguments{
\item{data_event}{Un data frame que contiene los datos de una
enfermedad o evento}
}
\value{
Un data frame que contiene los códigos geográficos estandarizados
de los datos de una enfermedad o evento
}
\description{
Función que estandariza los códigos geográficos de los datos
de una enfermedad o evento
}
\examples{
data(dengue2020)
data_event <- dengue2020
data_event <- limpiar_data_sivigila(data_event)
estandarizar_geo_cods(data_event = data_event)
}
