% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{filter_disease}
\alias{filter_disease}
\title{Filter Disease}
\usage{
filter_disease(name_disease, sivigila_summary_data)
}
\arguments{
\item{name_disease}{Name of the disease}

\item{data}{Data set}
}
\value{
The filtered data with the disease selected
}
\description{
Función que filtra en un conjunto de datos por el nombre de la enfermedad
Function that filters in a dataset by the disease name
}
\examples{
filter_disease(name_disease = "MALARIA", sivigila_summary_data)
}
