% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{filter_disease}
\alias{filter_disease}
\title{Filter disease}
\usage{
filter_disease(name_disease, sivigila_summary_data)
}
\arguments{
\item{name_disease}{Name of the disease}

\item{sivigila_summary_data}{The data set}
}
\value{
Data filtered with the disease selected
}
\description{
Function that filters by disease name in a dataset
}
\examples{
sivigila_summary_data <- import_sivigila_summary_data()
sivigila_summary_data <- clean_header(sivigila_summary_data)
filter_disease("MALAR", sivigila_summary_data)
}
