% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{format_dates_values}
\alias{format_dates_values}
\title{format_dates_values}
\usage{
format_dates_values(
  disease_data,
  date_format = "\%AAAA-\%MM-\%DD",
  col_names = c()
)
}
\arguments{
\item{disease_data}{Disease data}

\item{date_format}{Date format}

\item{col_names}{Data set column names}
}
\value{
The data with formatted dates
}
\description{
Función que da un formato especifico a una fecha
Function that gives a specific format to a date
}
\examples{
disease_data <- import_data_disease_by_year(2020, "DENGUE")
remove_error_dates(disease_data, date_format = "\%AAAA-\%MM-\%DD", col_names = c("INI_SIN", "FEC_HOS"))
}
