% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{geo_filtro}
\alias{geo_filtro}
\title{Filtrar por departamentos y municipios}
\usage{
geo_filtro(data_event, nombre_dpto = NULL, nombre_mun = NULL)
}
\arguments{
\item{data_event}{Un data frame con los datos de una enfermedad o evento}

\item{nombre_dpto}{Un character (cadena de caracteres) que contiene
el nombre del departamento; valor por defecto NULL}

\item{nombre_mun}{Un character (cadena de caracteres) que contiene el
nombre del municipio; su valor por defecto es NULL}
}
\value{
Un data frame con los datos filtrados con la enfermedad,
departamentos y municipios seleccionados
}
\description{
Función que filtra los datos de una enfermedad o evento por departamentos
y municipios
}
\examples{
data(dengue2020)
data_event <- dengue2020
data_event <- limpiar_data_sivigila(data_event, 2020)
geo_filtro(data_event, nombre_dpto = "ANTIOQUIA")
}
