% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{geo_filtro}
\alias{geo_filtro}
\title{Filtrar por departamentos y municipios}
\usage{
geo_filtro(data_event, nombre_dpto = NULL, nombre_mun = NULL)
}
\arguments{
\item{data_event}{Los datos de una enfermedad o evento}

\item{nombre_dpto}{El nombre del departamento}

\item{nombre_mun}{El nombre del municipio}
}
\value{
Datos filtrados con la enfermedad, departamentos y
municipios seleccionados
}
\description{
Función que filtra los datos de una enfermedad o evento por departamentos
y municipios
}
\examples{
data_event <- import_data_event(2019, "DENGUE")
data_event <- limpiar_encabezado(data_event)
geo_filtro(data_event, nombre_dpto = "ANTIOQUIA")
}
