% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{get_depto_codes}
\alias{get_depto_codes}
\title{Get Department Codes}
\usage{
get_depto_codes(geo_codes)
}
\arguments{
\item{deptos_data}{Departments data}
}
\value{
The Departments data with code and name
}
\description{
Función que obtiene el listado de departamentos de Colombia con su nombre y código
Function that gets the list of departments of Colombia with their name and code
}
\examples{
get_depto_codes(geo_codes)
}
