% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{get_depto_codes}
\alias{get_depto_codes}
\title{Get department codes}
\usage{
get_depto_codes(geo_codes)
}
\arguments{
\item{geo_codes}{Geographical codes (Colombia departments and municipalities)}
}
\value{
The Departments data with code and name
}
\description{
Function that gets the list of departments of Colombia with their name and code
}
\examples{
geo_codes <- import_geo_codes()
get_depto_codes(geo_codes)
}
