% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_depto_names}
\alias{get_depto_names}
\title{Get department names}
\usage{
get_depto_names(disease_data)
}
\arguments{
\item{disease_data}{The disease data}
}
\value{
Dataframe with the department names
}
\description{
Function that gets the department names
}
\examples{
disease_data <- import_linelist_disease_year(2020, "DENGUE")
}
