% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_info_depts}
\alias{get_info_depts}
\title{Get geographic information of the disease data}
\usage{
get_info_depts(department = NULL, municipalitie = NULL)
}
\arguments{
\item{department}{The disease data}

\item{municipalitie}{The disease data}
}
\value{
The geographic codes of the disease data
}
\description{
Function that standardizes the geographic codes of the disease data
}
\examples{
get_info_depts(department = "ANTIOQUIA")
}
