% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_months_major_cases}
\alias{get_months_major_cases}
\title{Get months major cases}
\usage{
get_months_major_cases(
  disease_data,
  col_dates,
  col_cases = "casos",
  top = 3,
  concat_values = TRUE
)
}
\arguments{
\item{disease_data}{The disease data}

\item{col_dates}{Column names in the disease data that contains the dates}

\item{col_cases}{Column name in the disease data that contains the cases number}

\item{top}{Top}

\item{concat_values}{Indicates if it is required to concatenate the months as a string}
}
\value{
Months major cases
}
\description{
Function that gets the months with the major number of cases
}
\examples{
disease_data <- import_data_disease_by_year(2020, "DENGUE")
get_months_major_cases(disease_data, col_dates, col_cases, top = 3, concat_values = TRUE)
}
