% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{get_name_file_path}
\alias{get_name_file_path}
\title{Get name file path or URL}
\usage{
get_name_file_path(path)
}
\arguments{
\item{path}{The path or URL}
}
\value{
The name file for download
}
\description{
Function that gets the file name for download from a URL or path
}
\examples{
get_name_file_path("DENGUE")
}
