% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping_data.R
\name{get_path_data_disease_by_year}
\alias{get_path_data_disease_by_year}
\title{Get Path Data Disease by Year}
\usage{
get_path_data_disease_by_year(year, disease_name)
}
\arguments{
\item{year}{Specific year}

\item{disease_name}{Disease name}
}
\value{
Path or URL for download the disease data by a specific year
}
\description{
Función que obtiene la ruta o URL del servicio de SIVIGILA para descargar los datos de una enfermedad por un año especifico\n
Function that gets the path or URL of the SIVIGILA service to download the data of a disease for a specific year
}
\examples{
get_path_data_disease_by_year(year = 2010, disease_name = "DENGUE")
}
