% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping_data.R
\name{get_path_data_disease_by_year}
\alias{get_path_data_disease_by_year}
\title{Gets the download path for a disease by specific year}
\usage{
get_path_data_disease_by_year(year, disease_name)
}
\arguments{
\item{year}{Specific year to download}

\item{disease_name}{Disease name to download}
}
\value{
Path or URL for download the disease data by a specific year of SIVIGILA source
}
\description{
Function that gets the path or URL of the SIVIGILA service to download the data of a disease for a specific year
}
\examples{
get_path_data_disease_by_year(year = 2010, disease_name = "DENGUE")
}
