% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_by_age_and_gender}
\alias{group_by_age_and_gender}
\title{Group by ages, gender and cases}
\usage{
group_by_age_and_gender(
  disease_data,
  col_names = c("edad", "sexo"),
  percentage = TRUE
)
}
\arguments{
\item{disease_data}{The disease data}

\item{col_names}{Column names in the disease data that contains the ages and gender}

\item{percentage}{Indicates if it is required to add a percentage of cases as a column}
}
\value{
The disease data grouped by ages, gender and number cases
}
\description{
Function that groups the data by age, gender and cases number
}
\examples{
disease_data <- import_data_disease_by_year(2019, "DENGUE")
group_by_age_and_gender(disease_data, col_names = c("edad", "sexo"), percentage = TRUE)
}
