% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_by_age_range_and_cases}
\alias{group_by_age_range_and_cases}
\title{Group By Age Range and Cases}
\usage{
group_by_age_range_and_cases(
  disease_data,
  col_name,
  var_a = NULL,
  min_val,
  max_val,
  step
)
}
\arguments{
\item{disease_data}{Disease data}

\item{col_name}{Colunm name}

\item{var_a}{Additional variables}

\item{min_val}{Min value to group}

\item{max_val}{Max value to group}

\item{step}{Step for the range}
}
\value{
The disease data grouped by the age range and cases
}
\description{
Función que agrupa los datos por un rango de edades y por el número de casos
Function that groups the data by an age range and cases number
}
\examples{
disease_data <- import_data_disease_by_year(2019, "DENGUE")
disease_dt_by_age <- group_by_columns_and_cases(disease_data, c("edad", "semana"), 
wt_percentage = TRUE)
group_by_age_range_and_cases(disease_dt_by_age, "edad", min_val = 0, 
max_val = max(disease_dt_by_age$edad), step = 10)
}
