% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_by_age_range_and_cases}
\alias{group_by_age_range_and_cases}
\title{Group By Age Range and Cases}
\usage{
group_by_age_range_and_cases(
  disease_data,
  col_name,
  var_a = NULL,
  min_val,
  max_val,
  step
)
}
\arguments{
\item{disease_data}{Disease data}

\item{col_name}{Colunm name}

\item{min_val}{Min value to group}

\item{max_val}{Max value to group}

\item{step}{Step for the range}
}
\value{
The disease data grouped by the age range and cases
}
\description{
Función que agrupa los datos por un rango de edades y por el número de casos
Function that groups the data by an age range and cases number
}
\examples{
group_by_age_range_and_cases(disease_data, min_val = 0, max_val = 100, step = 5)
}
