% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_by_columns_and_cases}
\alias{group_by_columns_and_cases}
\title{Group by columns and cases}
\usage{
group_by_columns_and_cases(disease_data, col_names, wt_percentage = FALSE)

group_by_columns_and_cases(disease_data, col_names, wt_percentage = FALSE)
}
\arguments{
\item{disease_data}{The disease data}

\item{col_names}{The column name(s)}

\item{wt_percentage}{Indicates if it is required to add a percentage of cases as a column}
}
\value{
The disease data grouped by a specific column name(s) and cases number

The disease data grouped by a specific column name(s) and cases number
}
\description{
Function that groups the disease data by a specific column name(s) and cases number

Function that groups the disease data by a specific column name(s) and cases
}
\examples{
disease_data <- import_data_disease_by_year(2019, "DENGUE")
group_by_columns_and_cases(disease_data, col_names = "sexo", wt_percentage = TRUE)
group_by_columns_and_cases(disease_data, col_names = c("sexo", "semana"))
disease_data <- import_data_disease_by_year(2019, "DENGUE")
group_by_columns_and_cases(disease_data, col_names = "sexo", wt_percentage = TRUE)
group_by_columns_and_cases(disease_data, col_names = c("sexo", "semana"))
}
