% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_by_columns_and_cases}
\alias{group_by_columns_and_cases}
\title{Group By Columns and Cases}
\usage{
group_by_columns_and_cases(disease_data, col_names, wt_percentage = FALSE)
}
\arguments{
\item{disease_data}{Disease data}

\item{col_names}{The column name(s)}

\item{wt_percentage}{Indicates if it is required to add a percentage of cases as a column}
}
\value{
The disease data grouped by a specific column name(s) and cases
}
\description{
Función que agrupa los datos por el nombre de las columna(s) y los casos\n
Function that groups the data by a specific column name and cases
}
\examples{
group_by_columns_and_cases(disease_data, col_names = "SEXO", wt_percentage = TRUE)
group_by_columns_and_cases(disease_data, col_names = c("SEXO","SEMANA"))
}
