% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_by_gender}
\alias{group_by_gender}
\title{group_by_gender}
\usage{
group_by_gender(disease_data, col_name = "sexo", percentage = T)
}
\arguments{
\item{disease_data}{Disease data}

\item{col_name}{Data set column name}
}
\value{
The disease data grouped by gender date and cases
}
\description{
Función que agrupa los datos por genero y los casos
Function that groups the data by gender and cases
}
\examples{
disease_data <- import_data_disease_by_year(2019, "DENGUE")
group_by_gender(disease_data, col_name = "sexo", )
}
