% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_by_gender_and_week}
\alias{group_by_gender_and_week}
\title{Group by gender, epidemiological week and cases}
\usage{
group_by_gender_and_week(
  disease_data,
  col_names = c("sexo", "semana"),
  percentage = TRUE
)
}
\arguments{
\item{disease_data}{The disease data}

\item{col_names}{Column names in the disease data that contains the gender and the epidemiological weeks}

\item{percentage}{Indicates if it is required to add a percentage of cases as a column}
}
\value{
The disease data grouped by gender, epidemiological week and cases number
}
\description{
Function that groups the disease data by gender, epidemiological week and cases number
}
\examples{
disease_data <- import_data_disease_by_year(2019, "DENGUE")
group_by_gender_and_week(disease_data, col_names = c("sexo", "semana"), percentage = TRUE)
}
