% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{group_epiweek_cases}
\alias{group_epiweek_cases}
\title{Group By week and cases}
\usage{
group_epiweek_cases(disease_data)
}
\arguments{
\item{disease_data}{The disease data}
}
\value{
The disease data grouped by epidemiological week 
and cases number
}
\description{
Function that groups the disease data by epidemiological week 
and cases number
}
\examples{
disease_data <- import_linelist_disease_year(2019, "DENGUE")
disease_data <- clean_header(disease_data)
group_epiweek_cases(disease_data)
}
