% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data_canal_endemico}
\alias{import_data_canal_endemico}
\title{Importar datos de una enfermedad o evento para construir
su canal endémico}
\usage{
import_data_canal_endemico(nombre_event, year)
}
\arguments{
\item{nombre_event}{Un character (cadena de caracteres) que
contiene el nombre de la enfermedad o evento}

\item{year}{Un numeric (numerico) que contiene el año
de referencia para la descarga de los datos}
}
\value{
Un `data.frame` con los datos de los últimos cinco años
de una enfermedad
}
\description{
Función que importa los datos necesarios de una enfermedad
o evento para construir el canal endémico desde la fuente
de SIVIGILA
}
\examples{
import_data_canal_endemico(nombre_event = "MALARIA",
                           year = 2020)
}
