% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data_disease_by_year}
\alias{import_data_disease_by_year}
\title{Import data of a disease by year from the SIVIGILA microdata source}
\usage{
import_data_disease_by_year(year, disease_name, cache = TRUE)
}
\arguments{
\item{year}{The selected year}

\item{disease_name}{The disease name}

\item{cache}{Indicates if the downloaded data to be cached}
}
\value{
The disease data by year from the SIVIGILA microdata source
}
\description{
Function that obtains the data of a disease by year
}
\examples{
import_data_disease_by_year(2018, "DENGUE")
}
