% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data_endemic_channel}
\alias{import_data_endemic_channel}
\title{Import data of a disease to build its endemic channel}
\usage{
import_data_endemic_channel(disease_name, year)
}
\arguments{
\item{disease_name}{The disease name}

\item{year}{Last year}
}
\value{
The last five years data of a disease
}
\description{
Function that imports the last five years of a disease data for building 
the endemic channel from SIVIGILA source
}
\examples{
import_data_endemic_channel("MALARIA", 2020)
}
