% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data_event}
\alias{import_data_event}
\title{Importar los datos de una enfermedad o evento por año
desde los microdatos de SIVIGILA}
\usage{
import_data_event(year, nombre_event, cache = TRUE)
}
\arguments{
\item{year}{Un numeric (numerico) con el año deseado para la descarga
de los datos}

\item{nombre_event}{Un character (cadena de caracteres) con el nombre de
la enfermedad o evento}

\item{cache}{Un boolean (TRUE o FALSE) que indica si los datos descargados
deben ser almacenados en caché; su valor por defecto es TRUE}
}
\value{
Un `data.frame` con los datos de la enfermedad o evento seleccionado
por año desde los microdatos de SIVIGILA
}
\description{
Función que obtiene los datos de una enfermedad por año
desde los microdatos de SIVIGILA
}
\examples{
import_data_event(2018, "DENGUE")
}
