% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data_event}
\alias{import_data_event}
\title{Importar los datos de una enfermedad o evento por año
desde los microdatos del SIVIGILA}
\usage{
import_data_event(
  nombre_event,
  years,
  ruta_dir = NULL,
  cache = FALSE,
  consentimiento = NULL
)
}
\arguments{
\item{nombre_event}{Un `character` (cadena de caracteres) con el nombre de
la enfermedad o evento.}

\item{years}{Un `numeric` (númerico) con el año o años deseado(s) para
la descarga de los datos.}

\item{ruta_dir}{Un `character` (cadena de caracteres) que contiene la ruta
del directorio donde se almacenaran los datos del evento o enfermedad. Si
no se proporciona ningún valor en este parámetro, el usuario debe dar su
consentimiento para que los datos sean almacenados temporalmente en el
directorio de usuario de R (`R user dir`). Su valor por defecto es `NULL`.}

\item{cache}{Un `logical` (`TRUE` o `FALSE`) que indica si los datos.
descargados deben ser almacenados en caché; su valor por defecto es `FALSE`.}

\item{consentimiento}{Un `character` (cadena de caracteres) que indica si
el usuario acepta almacenar los datos de la enfermedad o evento en el
directorio del paquete `extdata`. Esto es útil si se desea que los
datos se almacenen en caché y no se ha proporcionado ningún valor en el
parámetro `ruta_dir`. Los valores permitidos son `"Si"` o `"No"`. Si se
proporciona `"Si"`, los datos se guardarán en el directorio del paquete.
Si se proporciona `"No"`, los datos no se guardarán y se solicitará
especificar un valor para el parámetro `ruta_dir`. Su valor por defecto
es `NULL`.}
}
\value{
Un `data.frame` con los datos del año de la enfermedad o evento
seleccionado desde los microdatos del SIVIGILA.
}
\description{
Función que importa los datos de una enfermedad o evento por
año desde los microdatos del SIVIGILA.
}
\examples{
\donttest{
import_data_event(nombre_event = "DENGUE",
                  years = 2020,
                  consentimiento = "Si")
}
\donttest{
import_data_event(nombre_event = "ZIKA",
                  years = c(2019, 2020),
                  consentimiento = "Si",
                  cache = TRUE)
import_data_event(nombre_event = "ZIKA",
                  years = seq(2018, 2020),
                  consentimiento = "Si")
}
}
