% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data_event}
\alias{import_data_event}
\title{Importar los datos de una enfermedad o evento por año
desde los microdatos del SIVIGILA}
\usage{
import_data_event(nombre_event, years, cache = TRUE)
}
\arguments{
\item{nombre_event}{Un `character` (cadena de caracteres) con el nombre de
la enfermedad o evento}

\item{years}{Un `numeric` (numerico) con el año o años deseado(s) para
la descarga de los datos}

\item{cache}{Un `boolean` (TRUE o FALSE) que indica si los datos descargados
deben ser almacenados en caché; su valor por defecto es `TRUE`}
}
\value{
Un `data.frame` con los datos del año de la enfermedad o evento
seleccionado desde los microdatos del SIVIGILA
}
\description{
Función que importa los datos de una enfermedad o evento por año
desde los microdatos del SIVIGILA
}
\examples{
\donttest{
import_data_event(nombre_event = "DENGUE",
                  years = 2020,
                  cache = TRUE)
}
\dontrun{
import_data_event(nombre_event = "DENGUE",
                  years = c(2019, 2020),
                  cache = TRUE)
import_data_event(nombre_event = "DENGUE",
                  years = seq(2007, 2020),
                  cache = TRUE)
}
}
