% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_geo_cods}
\alias{import_geo_cods}
\title{Importar datos geográficos de Colombia}
\usage{
import_geo_cods(url_data = NULL)
}
\arguments{
\item{url_data}{Un `character` (cadena de caracteres) que contiene
la URL de los datos geográficos; su valor por defecto es NULL}
}
\value{
Un `data.frame` con los nombres y códigos de los departamentos
y municipios de Colombia
}
\description{
Función que importa los nombres y códigos de los departamentos
y municipios de Colombia a través de una URL
}
\examples{
import_geo_cods(url_data =
"https://www.datos.gov.co/api/views/gdxc-w37w/rows.csv?accessType=DOWNLOAD")
}
