% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_geo_cods}
\alias{import_geo_cods}
\title{Importar datos geográficos de Colombia}
\usage{
import_geo_cods(descargar = FALSE)
}
\arguments{
\item{descargar}{Un `boolean` (`TRUE` o `FALSE`) que indica si los
datos se deben descargar desde la API de datos abiertos; su valor por
defecto es `FALSE`}
}
\value{
Un `data.frame` con los nombres y códigos de los departamentos
y municipios de Colombia
}
\description{
Función que importa los nombres y códigos de
los departamentos y municipios de Colombia a través de una URL
}
\examples{
import_geo_cods(descargar = FALSE)
}
