% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_pob_proyecciones}
\alias{import_pob_proyecciones}
\title{Importar las proyecciones DANE del a침o 2005 hasta el 2035}
\usage{
import_pob_proyecciones(year)
}
\arguments{
\item{year}{Un `numeric` (numerico) con el a침o a importar de las
proyecciones poblacionales DANE.}
}
\value{
Un `data.frame` con las proyecciones poblacionales DANE.
}
\description{
Funci칩n que importa las proyecciones poblacionales
DANE desde el a침o 2005 hasta el 2035.
}
\examples{
\donttest{
import_pob_proyecciones(year = 2020)
}
}
