% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_pob_proyecciones}
\alias{import_pob_proyecciones}
\title{Importar las proyecciones DANE del año 2005 hasta el 2035}
\usage{
import_pob_proyecciones()
}
\value{
Un `data.frame` con las proyecciones poblacionales DANE
}
\description{
Función que importa las proyecciones poblacionales DANE desde
el año 2005 hasta el 2035
}
\examples{
\dontrun{
import_pob_proyecciones()
}
}
