% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_pob_riesgo}
\alias{import_pob_riesgo}
\title{Importar la población a riesgo de un evento o enfermedad}
\usage{
import_pob_riesgo(event, year)
}
\arguments{
\item{event}{Un `character` (cadena de caracteres) o un `numeric` (numerico)
con el nombre o código de la enfermedad o evento}

\item{year}{Un `numeric` (numerico) con el año deseado de la población a
riesgo}
}
\value{
Un `data.frame` con la población a riesgo de un año específico
}
\description{
Función que importa la población a riesgo de un evento o enfermedad de un
año específico
}
\examples{
\donttest{
import_pob_riesgo(event = "Dengue", year = 2020)
}
}
