% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_sep_data}
\alias{import_sep_data}
\title{Importar datos con un separador específico}
\usage{
import_sep_data(ruta_data = NULL, ruta_dir = NULL, cache = FALSE)
}
\arguments{
\item{ruta_data}{Un `character` (cadena de caracteres) que contiene
la URL de los datos de SIVIGILA.}

\item{ruta_dir}{Un `character` (cadena de caracteres) que contiene la ruta
del directorio donde se almacenarán los datos del evento o enfermedad.
Su valor por defecto es `NULL`.}

\item{cache}{Un `logical` (`TRUE` o `FALSE`) que indica si los datos
descargados deben ser almacenados en caché. Su valor por defecto
es `FALSE`.}
}
\value{
Un `data.frame` con los datos tabulados.
}
\description{
Función que importa e identifica el separador que tiene los
datos para tabularlos.
}
\keyword{internal}
