% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_sep_data}
\alias{import_sep_data}
\title{Importar datos con un separador específico}
\usage{
import_sep_data(path_data)
}
\arguments{
\item{path_data}{Un character (cadena de caracteres) que contiene
la URL de los datos de SIVIGILA}
}
\value{
Un `data.frame` con los datos
}
\description{
Función que identifica el separador que tiene
los datos para tabularla
}
\examples{
import_sep_data(path_data =
"https://www.datos.gov.co/api/views/qvnt-2igj/rows.csv?accessType=DOWNLOAD")
}
