% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_shape_map}
\alias{import_shape_map}
\title{Importar el Shapefile del mapa de Colombia}
\usage{
import_shape_map()
}
\value{
Un objeto `sf` que contiene los elementos del Shapefile
del mapa.
}
\description{
Función que importa el Shapefile del mapa de Colombia.
}
\keyword{internal}
