% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_sivigila_summary_data}
\alias{import_sivigila_summary_data}
\title{Import SIVIGILA summary data}
\usage{
import_sivigila_summary_data(
  url_data = config::get(file = system.file("extdata", "config.yml", package =
    "sivirep"), "sivigila_open_data_path")
)
}
\arguments{
\item{url_data}{URL of SIVIGILA data}
}
\value{
The data downloaded in csv format
}
\description{
Function that imports SIVIGILA data through a URL
}
\examples{
import_sivigila_summary_data(
  "https://www.datos.gov.co/api/views/qvnt-2igj/rows.csv?accessType=DOWNLOAD"
)
}
