% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{limpiar_data_sivigila}
\alias{limpiar_data_sivigila}
\title{Limpiar datos de SIVIGILA}
\usage{
limpiar_data_sivigila(data_event, year)
}
\arguments{
\item{data_event}{Los datos de una enfermedad o evento}

\item{year}{Año de los datos de una enfermedad o evento}
}
\value{
Datos limpios de la enfermedad
}
\description{
Función que limpia los datos seleccionados de una enfermedad o evento de
la fuente de SIVIGILA
}
\examples{
year <- 2019
data_event <- import_data_event(year, "DENGUE")
limpiar_data_sivigila(data_event, year)
}
