% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{limpiar_data_sivigila}
\alias{limpiar_data_sivigila}
\title{Limpiar datos de SIVIGILA}
\usage{
limpiar_data_sivigila(data_event)
}
\arguments{
\item{data_event}{Un data frame que contiene los datos de
una enfermedad o evento}
}
\value{
Un data frame con los datos limpios de la enfermedad o evento
}
\description{
Función que limpia los datos seleccionados de una enfermedad o evento de
la fuente de SIVIGILA
}
\examples{
data(dengue2020)
limpiar_data_sivigila(data_event = dengue2020)
}
