% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{limpiar_encabezado}
\alias{limpiar_encabezado}
\title{Limpiar las etiquetas del encabezado}
\usage{
limpiar_encabezado(data_event)
}
\arguments{
\item{data_event}{Un `data.frame` que contiene los datos de una
enfermedad o evento.}
}
\value{
Un `data.frame` con las etiquetas del encabezado formateadas
con guión bajo (_).
}
\description{
Función que limpia las etiquetas del encabezado de los
datos de una enfermedad o evento.
}
\examples{
data(dengue2020)
limpiar_encabezado(data_event = dengue2020)
}
