% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{limpiar_val_atipic}
\alias{limpiar_val_atipic}
\title{Limpiar las edades de los datos de una enfermedad o evento}
\usage{
limpiar_val_atipic(data_event)
}
\arguments{
\item{data_event}{Los datos de una enfermedad o evento
de una enfermedad o evento que contiene las edades}
}
\value{
Los datos de una enfermedad o evento con las edades limpias
}
\description{
Función que limpia las edades de los datos de una enfermedad o evento
}
\examples{
data_event <- import_data_event(2020, "DENGUE")
data_event <- limpiar_encabezado(data_event)
}
