% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{limpiar_val_atipic}
\alias{limpiar_val_atipic}
\title{Limpiar los valores atípicos de los datos}
\usage{
limpiar_val_atipic(data_event)
}
\arguments{
\item{data_event}{Un `data.frame` que contiene los datos de una
enfermedad o evento.}
}
\value{
Un `data.frame` con los datos de una enfermedad o
evento con los valores atípicos limpios (convertidos a `NA`).
}
\description{
Función que limpia los valores atípicos de los datos
de una enfermedad o evento del SIVIGILA.
}
\examples{
data(dengue2020)
data_limpia <- limpiar_encabezado(data_event = dengue2020)
limpiar_val_atipic(data_limpia)
}
