% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{list_events}
\alias{list_events}
\title{Importar una enfermedad o evento y a침os disponibles para descargar
desde los microdatos de SIVIGILA}
\usage{
list_events()
}
\value{
Las una enfermedad o evento y los a침os disponibles desde
los microdatos de SIVIGILA
}
\description{
Funci칩n que obtiene las una enfermedad o evento y los a침os disponibles
desde los microdatos de SIVIGILA
}
\examples{
list_events()
}
