% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{list_events}
\alias{list_events}
\title{Importar las enfermedades y a침os disponibles disposibles
para su descarga desde los microdatos de SIVIGILA}
\usage{
list_events()
}
\value{
Una list con las enfermedades y los a침os disponibles
para su descarga desde los microdatos de SIVIGILA
}
\description{
Funci칩n que obtiene las enfermedades y los a침os disponibles
para su descarga desde los microdatos de SIVIGILA
}
\examples{
\donttest{
list_events()
}
}
