% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_data.R
\name{obtener_casos_pob_especial}
\alias{obtener_casos_pob_especial}
\title{Obtener poblaci贸n especial y casos}
\usage{
obtener_casos_pob_especial(data_event)
}
\arguments{
\item{data_event}{Los datos de una enfermedad o evento}
}
\value{
Los casos por tipo de poblaci贸n especial de una enfermedad
}
\description{
Funci贸n que obtiene los casos por tipo de poblaci贸n
especial de una enfermedad
}
\examples{
data_event <- import_data_event(2019, "DENGUE")
obtener_casos_pob_especial(data_event)
}
