% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtener_dptos}
\alias{obtener_dptos}
\title{Obtener departamentos de Colombia}
\usage{
obtener_dptos()
}
\value{
Un `data.frame` con los departamentos de Colombia.
}
\description{
Función que obtiene los departamentos de Colombia.
}
\examples{
obtener_dptos()
}
