% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_sivirep.R
\name{obtener_estetica_escala}
\alias{obtener_estetica_escala}
\title{Obtener la estetica de una escala para un grafico de \pkg{sivirep}}
\usage{
obtener_estetica_escala(escala = 0, nombre, etiquetas = NULL)
}
\arguments{
\item{escala}{Un `numeric` (numerico) que indica la cantidad de valores que
contiene la escala.}

\item{nombre}{Un `character` (cadena de caracteres) que contiene el nombre
de la escala.}

\item{etiquetas}{Un `character` (cadena de caracteres) que contiene las
etiquetas de la escala.}
}
\value{
Un objeto `scale_fill_manual` de \pkg{ggplot2}.
}
\description{
Función que genera la estética de una escala para un gráfico
de \pkg{sivirep}.
}
\keyword{internal}
