% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtener_eventos_relacionados}
\alias{obtener_eventos_relacionados}
\title{Obtener los eventos relacionados}
\usage{
obtener_eventos_relacionados(nombre_event, years)
}
\arguments{
\item{nombre_event}{Un `character` (cadena de caracteres) con el
nombre de la enfermedad o evento}

\item{years}{Un `numeric` (numerico) el a침o  o a침os deseados para
la descarga de los datos}
}
\value{
Un `array` con los eventos relacionados por a침o desde
los microdatos de SIVIGILA
}
\description{
Funci칩n que obtiene los eventos relacionados o tipos de un evento
principal
}
\keyword{internal}
