% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtener_info_depts}
\alias{obtener_info_depts}
\title{Obtener información geográfica de los datos de la enfermedad o evento}
\usage{
obtener_info_depts(dpto = NULL, munpio = NULL)
}
\arguments{
\item{dpto}{Un character (cadena de caracteres) que contiene el nombre
del departamento; su valor por defecto es NULL}

\item{munpio}{Un character (cadena de caracteres) que contiene los datos del
municipio; su valor por defecto es NULL}
}
\value{
Un data frame con la información geográfica de los datos de
la enfermedad o evento
}
\description{
Función que obtiene la información geográfica de los datos de la enfermedad
o evento
}
\examples{
obtener_info_depts(dpto = "ANTIOQUIA")
}
