% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtener_info_depts}
\alias{obtener_info_depts}
\title{Obtener información geográfica de los datos de la enfermedad o evento}
\usage{
obtener_info_depts(dpto = NULL, mpio = NULL)
}
\arguments{
\item{dpto}{Un `character` (cadena de caracteres) o `numeric` (numérico)
que contiene el nombre o código del departamento; su valor por defecto
es `NULL`.}

\item{mpio}{Un `character` (cadena de caracteres) o `numeric` (numérico)
que contiene el nombre o código del municipio; su valor por defecto
es `NULL`.}
}
\value{
Un `data.frame` con la información geográfica de los datos de
la enfermedad o evento.
}
\description{
Función que obtiene la información geográfica de los datos de
la enfermedad o evento.
}
\examples{
obtener_info_depts(dpto = "ANTIOQUIA")
obtener_info_depts(dpto = "ANTIOQUIA", mpio = "MEDELLIN")
obtener_info_depts(dpto = "05")
obtener_info_depts(dpto = "05", mpio = "05001")
obtener_info_depts(dpto = 05, mpio = 05001)
obtener_info_depts(dpto = 05, mpio = 001)
obtener_info_depts(dpto = "bogota dc", mpio = "bogota dc")
}
