% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtener_nombres_muns}
\alias{obtener_nombres_muns}
\title{Obtener nombre del municipio en Colombia}
\usage{
obtener_nombres_muns(data_geo, code_dept, cod_mun)
}
\arguments{
\item{data_geo}{Un data frame que contiene los códigos
geográficos (departamentos y municipios de Colombia)}

\item{code_dept}{Un numeric (numerico) que contiene el código
del departamento}

\item{cod_mun}{Un numeric (numerico) que contiene el código
del municipio}
}
\value{
Un character (cadena de caracteres) con el nombre del municipio
}
\description{
Función que obtiene el nombre del municipio
}
\examples{
data_geo <- import_geo_cods()
obtener_nombres_muns(data_geo,
                     code_dept = "05",
                     cod_mun = "001")
}
