% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{obtener_ruta_descarga}
\alias{obtener_ruta_descarga}
\title{Obtener el nombre del archivo de ruta o URL}
\usage{
obtener_ruta_descarga(ruta)
}
\arguments{
\item{ruta}{La ruta o URL}
}
\value{
El nombre del archivo a descargar
}
\description{
Función que obtiene el nombre del archivo a descargar desde una URL o ruta
}
\examples{
obtener_ruta_descarga("DENGUE")
}
