% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtener_tip_ocurren_geo}
\alias{obtener_tip_ocurren_geo}
\title{Obtener columnas de ocurrencia geográfica de los datos de la enfermedad}
\usage{
obtener_tip_ocurren_geo(code_disease)
}
\arguments{
\item{code_disease}{El código de la enfermedad}
}
\value{
Las columnas de ocurrencia geográfica de los datos de la enfermedad
}
\description{
Función que obtiene las columnas de ocurrencia geográfica de los
datos de la enfermedad o evento
}
\examples{
obtener_tip_ocurren_geo(code_disease = 210)
}
