% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtener_tip_ocurren_geo}
\alias{obtener_tip_ocurren_geo}
\title{Obtener columnas de ocurrencia geográfica de los datos de la
enfermedad o evento}
\usage{
obtener_tip_ocurren_geo(cod_event = NULL, nombre_event = NULL)
}
\arguments{
\item{cod_event}{Un `numeric` (numérico) o `character`
(cadena de caracteres) que contiene el código de la enfermedad o evento.}

\item{nombre_event}{Un `character` (cadena de caracteres) con el nombre de
la enfermedad o evento.}
}
\value{
Un `data.frame` con las columnas de ocurrencia geográfica de los
datos de la enfermedad o evento.
}
\description{
Función que obtiene las columnas de ocurrencia geográfica
de los datos de la enfermedad o evento.
}
\examples{
obtener_tip_ocurren_geo(cod_event = 210)
}
