% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_age}
\alias{plot_age}
\title{Plot cases distribution by age}
\usage{
plot_age(data_grouped, col_name = "edad", percentage = FALSE)
}
\arguments{
\item{data_grouped}{The disease data grouped}

\item{col_name}{Column name in the disease data grouped that contains 
the ages}

\item{percentage}{Indicates if the data has percentages}
}
\value{
A plot of cases distribution by age
}
\description{
Function that generates the plot of cases distribution by age
}
\examples{
disease_data <- import_linelist_disease_year(2020, "DENGUE")
disease_data <- clean_header(disease_data)
data_grouped <- group_age(disease_data, 
                           col_name = "edad", 
                           percentage = FALSE)
plot_age(data_grouped, 
         col_name = "edad", 
         percentage = FALSE)
}
