% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_by_variable}
\alias{plot_by_variable}
\title{Plot by Variable}
\usage{
plot_by_variable(
  data,
  var_x,
  var_y,
  var_per = NULL,
  var_fill,
  wt_per = TRUE,
  label_x,
  label_y,
  scale_name = NULL,
  scale_labels = NULL,
  diagram_title,
  legend_pos,
  bar_wd = 1,
  text_sz,
  show_val = TRUE
)
}
\arguments{
\item{data}{The data to plot}

\item{var_x}{The variable name for axis x}

\item{var_y}{The variable name for axis y}

\item{var_fill}{The variable fill}

\item{wt_per}{Indicates if the data have percentages}

\item{label_x}{The label for axis x}

\item{label_y}{The label for axis y}

\item{scale_name}{The scale name}

\item{scale_labels}{The scale labels}

\item{diagram_title}{The diagram title}

\item{legend_pos}{The legend position}

\item{bar_wd}{The bar width}

\item{text_sz}{The text size}
}
\value{
The map for department
}
\description{
Función que genera gráficas por cualquier tipo de variable\n
Function that generates graphs by any type of variable
}
\examples{
plot_by_variable(disease_dt_by_gender_wk, var_x = "SEMANA", var_y = "Casos", var_fill = "SEXO", var_per = "Porcentaje", label_x = "Género", label_y = "Casos \n (Porcentajes)", scale_name = "Género", scale_labels = c("Femenino", "Masculino"), diagram_title = "Diagrama para la variable Sexo", legend_pos = "right", bar_wd = 1, text_sz = 0.5)
}
