% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_by_variable}
\alias{plot_by_variable}
\title{Plot by Variable}
\usage{
plot_by_variable(
  data,
  var_x,
  var_y,
  var_per = NULL,
  var_fill,
  wt_per = TRUE,
  label_x,
  label_y,
  scale_name = NULL,
  scale_labels = NULL,
  diagram_title,
  legend_pos,
  bar_wd = 1,
  text_sz,
  show_val = TRUE
)
}
\arguments{
\item{data}{The data to plot}

\item{var_x}{The variable name for axis x}

\item{var_y}{The variable name for axis y}

\item{var_per}{The variable name if the data have percentages}

\item{var_fill}{The variable fill}

\item{wt_per}{Indicates if the data have percentages}

\item{label_x}{The label for axis x}

\item{label_y}{The label for axis y}

\item{scale_name}{The scale name}

\item{scale_labels}{The scale labels}

\item{diagram_title}{The diagram title}

\item{legend_pos}{The legend position}

\item{bar_wd}{The bar width}

\item{text_sz}{The text size}

\item{show_val}{Indicates if the bars should show the values}
}
\value{
The map for department
}
\description{
Función que genera gráficas por cualquier tipo de variable
Function that generates graphs by any type of variable
}
\examples{
disease_data <- import_data_disease_by_year(2019, "DENGUE")
disease_dt_by_gender_wk <- group_by_columns_and_cases(disease_data, "SEXO", wt_percentage = TRUE)
plot_by_variable(disease_dt_by_gender_wk, var_x = "SEXO", var_y = "Casos", var_fill = "SEXO", 
var_per = "Porcentaje", label_x = "Gender", label_y = "Cases \n (Percentage)", 
scale_name = "Gender", scale_labels = c("Female", "Male"), diagram_title = "Cases by Gender", 
legend_pos = "right", bar_wd = 0.6, text_sz = 4)
}
