% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_edad}
\alias{plot_edad}
\title{Generar gráfico de distribución de casos por edad}
\usage{
plot_edad(data_agrupada, col_nombre = "edad", porcentaje = FALSE)
}
\arguments{
\item{data_agrupada}{Los datos de la enfermedad agrupados}

\item{col_nombre}{Nombre de la columna en los datos de la
enfermedad o evento agrupados que contiene las edades}

\item{porcentaje}{Indica si los datos tienen porcentajes}
}
\value{
Un gráfico de distribución de casos por edad
}
\description{
Función que genera el gráfico de distribución de casos por edad
}
\examples{
data_event <- import_data_event(2020, "DENGUE")
data_event <- limpiar_encabezado(data_event)
data_agrupada <- agrupar_edad(data_event,
                              col_nombre = "edad",
                              porcentaje = FALSE)
plot_edad(data_agrupada,
          col_nombre = "edad",
          porcentaje = FALSE)
}
