% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_edad_sex}
\alias{plot_edad_sex}
\title{Generar gráfico de distribución de casos por edad y sexo}
\usage{
plot_edad_sex(
  data_agrupada,
  col_nombres = c("edad", "sexo"),
  porcentaje = FALSE
)
}
\arguments{
\item{data_agrupada}{Los datos de la enfermedad agrupados}

\item{col_nombres}{Nombres de columna en los datos de la
enfermedad o evento agrupados que contienen
las edades y las semanas epidemiológicas}

\item{porcentaje}{Indica si los datos tienen porcentajes}
}
\value{
Un gráfico de distribución de casos por edad y sexo
}
\description{
Función que genera el gráfico de distribución de casos por edad y sexo
}
\examples{
data_event <- import_data_event(2020, "DENGUE")
data_event <- limpiar_encabezado(data_event)
data_agrupada <- agrupar_edad_sex(data_event,
                                  col_nombres = c("edad", "sexo"),
                                  porcentaje = FALSE)
plot_edad_sex(data_agrupada,
              col_nombres = c("edad", "sexo"),
              porcentaje = FALSE)
}
