% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_edad_sex}
\alias{plot_edad_sex}
\title{Generar gráfico de distribución de casos por edad y sexo}
\usage{
plot_edad_sex(data_agrupada, nomb_cols = c("edad", "sexo"), fuente_data = NULL)
}
\arguments{
\item{data_agrupada}{Un `data.frame` que contiene los datos de la
enfermedad o evento agrupados}

\item{nomb_cols}{Un `array` (arreglo) de `character` (cadena de caracteres)
con los nombres de las columnas de los datos agrupados de la enfermedad o
evento que contienen las edades y las semanas epidemiológicas; su valor
por defecto es `c("edad", "sexo")`}

\item{fuente_data}{Un `character` (cadena de caracteres) que contiene la
leyenda o fuente de información de los datos; su valor por defecto es `NULL`}
}
\value{
Un `plot` o gráfico de distribución de casos por edad y sexo
}
\description{
Función que genera el gráfico de distribución de casos por edad y sexo
}
\examples{
data(dengue2020)
data_limpia <- limpiar_data_sivigila(dengue2020)
data_agrupada <- agrupar_edad_sex(data_event = data_limpia,
                                  nomb_cols = c("edad", "sexo"),
                                  porcentaje = FALSE)
plot_edad_sex(data_agrupada = data_agrupada,
              nomb_cols = c("edad", "sexo"))
}
