% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_epiweek}
\alias{plot_epiweek}
\title{Plot epiweek}
\usage{
plot_epiweek(
  dat,
  var_week,
  var_cases,
  year,
  type = "week",
  xlabel = "Semana epidemiológica",
  ylabel = "Número de casos por semana"
)
}
\arguments{
\item{var_week}{Name of variable that indicates the week in the data}

\item{var_cases}{Name of variable that indicates the cases number in the data}

\item{year}{The year of the data}

\item{type}{The type (week, date)}

\item{xlabel}{The label of x axis}

\item{ylabel}{The label of y axis}

\item{data}{Disease data}
}
\value{
The chart of Epidemiological Week
}
\description{
Función que genera el reporte de la Semana epidemiológica
Function that generates the report of the Epidemiological Week
}
\examples{
plot_epiweek(data, var_week = "SEMANA", var_cases = "conteo_casos", year = 2019, type = "date", xlabel = "Fecha de semana epidemiológica")
}
