% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_epiweek}
\alias{plot_epiweek}
\title{Plot epidemiological weeks}
\usage{
plot_epiweek(
  dat,
  col_week,
  col_cases,
  year,
  type = "week",
  xlabel = "Semana epidemiologica",
  ylabel = "Numero de casos por semana"
)
}
\arguments{
\item{dat}{The disease data}

\item{col_week}{Column name in the disease data that contains the epidemiological weeks}

\item{col_cases}{Column Name in the disease data that contains the cases number}

\item{year}{Year of the disease data}

\item{type}{Time unit for axis x (week and date)}

\item{xlabel}{Label of axis x}

\item{ylabel}{Label of axis y}
}
\value{
The graphic of epidemiological weeks
}
\description{
Function that generates the graphic of the epidemiological weeks
}
\examples{
sivigila_summary_data <- import_sivigila_summary_data()
filtered_data <- filter_disease("MALAR", sivigila_summary_data)
plot_epiweek(filtered_data, col_week = "semana", col_cases = "conteo_casos", year = 2019, type = "date", xlabel = "Fecha de semana epidemiologica")
}
