% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_fecha_inisintomas}
\alias{plot_fecha_inisintomas}
\title{Generar gráfico de distribución de casos por fecha de inicio de síntomas}
\usage{
plot_fecha_inisintomas(
  data_agrupada,
  col_nombre = "ini_sin",
  uni_marca = "month"
)
}
\arguments{
\item{data_agrupada}{Un data frame que contiene los datos de la enfermedad
o evento agrupados}

\item{col_nombre}{Un character (cadena de caracteres) que contiene el
nombre de la columna en los datos de la enfermedad o evento
agrupados que contiene las fechas de inicio de síntomas; su valor por
defecto es "ini_sin"}

\item{uni_marca}{Un character (cadena de caracteres) que contiene la unidad
de las marcas del gráfico ("day": día, "month": mes y "year": año);
su valor por defecto es "month"}
}
\value{
Un plot o gráfico de la distribución de casos por fecha de inicio
de síntomas
}
\description{
Función que genera el gráfico de distribución de casos
por fecha de inicio de síntomas
}
\examples{
data_event <- import_data_event(2020, "DENGUE")
data_event <- limpiar_encabezado(data_event)
data_agrupada <- agrupar_fecha_inisintomas(
                                     data_event,
                                     col_nombre = "ini_sin",
                                     tipo = "month")
plot_fecha_inisintomas(data_agrupada = data_agrupada,
                       col_nombre = "ini_sin",
                       uni_marca = "month")
}
