% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_fecha_notifica}
\alias{plot_fecha_notifica}
\title{Generar gráfico de distribución de casos por fecha de notificación}
\usage{
plot_fecha_notifica(
  data_agrupada,
  nomb_col = "fec_not",
  uni_marca = "semanaepi",
  fuente_data = NULL
)
}
\arguments{
\item{data_agrupada}{Un `data.frame` que contiene los datos de la
enfermedad o evento agrupados}

\item{nomb_col}{Un `character` (cadena de caracteres) que contiene el
nombre de la columna en los datos de la enfermedad o evento agrupados con
las fechas de notificación; su valor por defecto es `"fec_not"`}

\item{uni_marca}{Un `character` (cadena de caracteres) que contiene la unidad
de las marcas del gráfico (`"dia"`, `"semanaepi"`y `"mes"``);
su valor por defecto es `"semanaepi"`}

\item{fuente_data}{Un `character` (cadena de caracteres) que contiene la
leyenda o fuente de información de los datos; su valor por defecto es `NULL`}
}
\value{
Un `plot` o gráfico de distribución de casos por fecha de
notificación
}
\description{
Función que genera el gráfico de distribución de casos por
fecha de notificación
}
\examples{
data(dengue2020)
data_limpia <- limpiar_data_sivigila(dengue2020)
data_agrupada <- agrupar_fecha_notifica(data_event = data_limpia,
                                        nomb_col = "fec_not")
plot_fecha_notifica(data_agrupada = data_agrupada,
                    nomb_col = "fec_not",
                    uni_marca = "semanaepi")
}
