% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gender}
\alias{plot_gender}
\title{plot_gender}
\usage{
plot_gender(disease_data, col_name = "sexo", percentage = T)
}
\arguments{
\item{disease_data}{Disease data}

\item{col_name}{Data set column name}
}
\value{
A plot of cases distribution by gender
}
\description{
Función que genera la gráficas de distribucion de casos por genero
Function that generates the plot of cases distribution by gender
}
\examples{
disease_data <- import_data_disease_by_year(2020, "DENGUE")
plot_gender(disease_data, col_name = "sexo", percentage = T)
}
