% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_map}
\alias{plot_map}
\title{Plot department map}
\usage{
plot_map(
  data_grouped,
  col_name_lj = "id",
  caption_label = NULL,
  department = NULL,
  municipality = NULL
)
}
\arguments{
\item{data_grouped}{The disease data grouped by department and cases number}

\item{col_name_lj}{Column name to join with the shape file}

\item{caption_label}{Caption or information source of the disease data}

\item{department}{Department name}

\item{municipality}{Municipality name}
}
\value{
The map by departments or municipalities with the cases number of 
a specific disease
}
\description{
Function that generates the map by departments or municipalities with the 
cases number of a specific disease
}
\examples{
disease_data <- import_linelist_disease_year(2019, "DENGUE")
disease_data <- clean_header(disease_data)
department_spacial_data <- group_dept(disease_data)
plot_map(department_spacial_data,
   col_name_lj = "id",
   caption_label = "Fuente: SIVIGILA, Instituto Nacional de Salud, Colombia",
   deparment = "Antioquia",
   municipalitie = "Envigado")
}
