% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_map_dpto}
\alias{plot_map_dpto}
\title{Generar mapa por departamento}
\usage{
plot_map_dpto(data_agrupada, col_nombre_lj = "id", fuente_data = NULL)
}
\arguments{
\item{data_agrupada}{Un data frame que contiene los datos de la enfermedad
o evento agrupados por departamento y número de casos}

\item{col_nombre_lj}{Un character (cadena de caracteres) que contiene el
nombre de la columna para unir con el archivo de forma (shape file)}

\item{fuente_data}{Un character (cadena de caracteres) que contiene la
leyenda o fuente de información de los datos de la enfermedad o evento}
}
\value{
Un plot con el mapa por departamento con el número de casos de
la enfermedad o evento
}
\description{
Función que genera el mapa por departamento con el número de casos
de una enfermedad o evento
}
\examples{
\dontrun{
data(dengue2020)
data_event <- dengue2020
data_event <- limpiar_data_sivigila(data_event, 2020)
data_event <- estandarizar_geo_cods(data_event)
data_dpto <- agrupar_dpto(data_event)
plot_map_dpto(data_dpto,
   col_nombre_lj = "id",
   fuente_data = "Fuente: SIVIGILA, Instituto Nacional de Salud, Colombia")
}
}
