% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_muns}
\alias{plot_muns}
\title{Generar gráfico de distribución de casos por municipios}
\usage{
plot_muns(data_agrupada, col_nombre = "nombre", fuente_data = NULL)
}
\arguments{
\item{data_agrupada}{Un data frame que contiene los datos de la
enfermedad o evento agrupados}

\item{col_nombre}{Un character (cadena de carácteres) con el nombre de
la columna de los datos agrupados de la enfermedad o evento que contiene
los municipios; su valor por defecto es "nombre"}

\item{fuente_data}{Un character (cadena de caracteres) que contiene la
leyenda o fuente de información de los datos; su valor por defecto es NULL}
}
\value{
Un plot o gráfico de distribución de casos por municipios
}
\description{
Función que genera el gráfico de distribución de casos por municipios
}
\examples{
data(dengue2020)
data_event <- dengue2020
data_event <- limpiar_data_sivigila(data_event, 2020)
data_event <- estandarizar_geo_cods(data_event)
data_agrupada <- agrupar_mun(data_event,
                             dept_nombre = "Antioquia")
plot_muns(data_agrupada,
          col_nombre = "nombre")
}
