% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R, R/plotting.R
\name{plot_muns}
\alias{plot_muns}
\title{Plot cases distribution of municipalities}
\usage{
plot_muns(data_grouped, col_name = "nombre", percentage = FALSE)

plot_muns(data_grouped, col_name = "nombre", percentage = FALSE)
}
\arguments{
\item{data_grouped}{The disease data grouped}

\item{col_name}{Column names in the disease data grouped that contains
the municipalities}

\item{percentage}{Indicates if the data has percentages}
}
\value{
A plot of cases distribution by municipalities

A plot of cases distribution by municipalities
}
\description{
Function that generates the plot of cases distribution by municipalities

Function that generates the plot of cases distribution by municipalities
}
\examples{
disease_data <- import_linelist_disease_year(2020, "DENGUE")
disease_data <- clean_header(disease_data)
data_grouped <- group_municipalities(disease_data,
                                  department = "Antioquia")
plot_muns(data_grouped,
                    col_name = "nombre",
                    percentage = FALSE)
disease_data <- import_linelist_disease_year(2020, "DENGUE")
disease_data <- clean_header(disease_data)
data_grouped <- group_municipalities(disease_data,
                                  department = "Antioquia")
plot_muns(data_grouped,
                    col_name = "nombre",
                    percentage = FALSE)
}
