% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_notification_date}
\alias{plot_notification_date}
\title{Plot cases distribution by notification date}
\usage{
plot_notification_date(
  data_grouped,
  col_name = "fec_not",
  break_tick_date = "month"
)
}
\arguments{
\item{data_grouped}{The disease data grouped}

\item{col_name}{Column name in the disease data grouped that contains 
the notification dates}

\item{break_tick_date}{Tick unit (day, month and year)}
}
\value{
A plot of cases distribution by onset notification date
}
\description{
Function that generates the plot of cases distribution by notification date
}
\examples{
disease_data <- import_data_disease_by_year(2020, "DENGUE")
data_grouped <- group_notification_date(disease_data, 
                                         col_name = "fec_not", 
                                         break_tick_date = "month")
plot_notification_date(data_grouped, col_name = "fec_not", break_tick_date = "month")
}
