% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_notification_date}
\alias{plot_notification_date}
\title{plot_notification_date}
\usage{
plot_notification_date(disease_data, col_name = "fec_not", type = "month")
}
\arguments{
\item{disease_data}{Disease data}

\item{col_name}{Data set column name}

\item{type}{Time unit}

\item{year}{Year}
}
\value{
A plot of cases distribution by onset symptoms date
}
\description{
Función que genera la gráficas de distribucion de casos por fecha de inicio de sintomas
Function that generates the plot of cases distribution by onset symptoms date
}
\examples{
disease_data <-  import_data_disease_by_year(2020, "DENGUE")
plot_onset_symptoms(disease_data, col_name = "fec_not", type = "month")
}
