% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pob_especial}
\alias{plot_pob_especial}
\title{Generar gráfico de distribución de casos por población especial}
\usage{
plot_pob_especial(data_agrupada, col_nombre = "poblacion", porcentaje = FALSE)
}
\arguments{
\item{data_agrupada}{Un data frame que contiene los datos de la
enfermedad o evento agrupados}

\item{col_nombre}{Un character (cadena de carácteres) con el nombre de
la columna de los datos agrupados de la enfermedad o evento que contiene
las poblaciones especiales; su valor por defecto es "poblacion"}

\item{porcentaje}{Un boolean (TRUE/FALSE) que indica si los datos tienen
porcentajes; su valor por defecto es FALSE}
}
\value{
Un plot o gráfico de distribución de casos por población especial
}
\description{
Función que genera el gráfico de distribución de casos por población especial
}
\examples{
data_event <- import_data_event(2020, "DENGUE")
data_event <- limpiar_data_sivigila(data_event)
data_agrupada <- agrupar_pob_especial(data_event,
                         col_nombre = "poblacion",
                         porcentaje = TRUE)
plot_pob_especial(data_agrupada = data_agrupada,
                  col_nombre = "poblacion",
                  porcentaje = FALSE)
}
