% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_semanaepi}
\alias{plot_semanaepi}
\title{Plot de las semanas epidemiológicas}
\usage{
plot_semanaepi(
  data,
  col_semana,
  col_casos,
  year,
  tipo = "week",
  etiqueta_x = "Semana epidemiologica",
  etiqueta_y = "Numero de casos por semana"
)
}
\arguments{
\item{data}{Un data frame con los datos de la enfermedad
o evento}

\item{col_semana}{Un data frame con el nombre de la columna
de los datos de la enfermedad o evento que contiene las semanas
epidemiológicas}

\item{col_casos}{Un character (cadena de caracteres) que contiene
el nombre de la columna de los datos de la enfermedad o evento
que contiene el número de casos}

\item{year}{Un numeric (numerico) que contiene el año de los datos
de la enfermedad o evento}

\item{tipo}{Un character (cadena de caracteres) que contiene la unidad de
tiempo para el eje x ("week": semana y "date": fecha)}

\item{etiqueta_x}{Un character (cadena de caracteres) que contiene la
etiqueta del eje x}

\item{etiqueta_y}{Un character (cadena de caracteres) que contiene la
etiqueta del eje y}
}
\value{
Un plot o gráfico de las semanas epidemiológicas
}
\description{
Función que genera el gráfico de las semanas epidemiológicas
}
\examples{
data_resumida_sivigila <- import_data_resumen_sivigila()
data_resumida_sivigila <- limpiar_encabezado(data_resumida_sivigila)
data_filtrada <- filtrar_event("MALAR", data_resumida_sivigila)
plot_semanaepi(data = data_filtrada,
               col_semana = "semana",
               col_casos = "casos",
               year = 2019,
               tipo = "date",
               etiqueta_x = "Fecha de semana epidemiologica")
}
