% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_sex}
\alias{plot_sex}
\title{Plot cases distribution by sex}
\usage{
plot_sex(data_grouped, col_name = "sexo", percentage = TRUE)
}
\arguments{
\item{data_grouped}{The disease data grouped}

\item{col_name}{Column name in the disease data grouped that contains the sex}

\item{percentage}{Indicates if the data has percentages}
}
\value{
A plot of cases distribution by sex
}
\description{
Function that generates the plot of cases distribution by sex
}
\examples{
disease_data <- import_data_disease_by_year(2020, "DENGUE")
disease_data <- clean_header(disease_data)
data_grouped <- group_sex(disease_data, col_name = "sexo", percentage = TRUE)
plot_sex(data_grouped, col_name = "sexo", percentage = TRUE)
}
