% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_sex_epiweek}
\alias{plot_sex_epiweek}
\title{Plot cases distribution by sex and epidemiological week}
\usage{
plot_sex_epiweek(
  data_grouped,
  col_names = c("sexo", "semana"),
  percentage = FALSE
)
}
\arguments{
\item{data_grouped}{The disease data grouped}

\item{col_names}{Column names in the disease data grouped that contains the sex and the epidemiological weeks}
}
\value{
A plot of cases distribution by sex and epidemiological week
}
\description{
Function that generates the plot of cases distribution by sex and epidemiological week
}
\examples{
disease_data <- import_data_disease_by_year(2020, "DENGUE")
data_grouped <- group_sex_and_week(disease_data, col_names = c("sexo", "semana"), percentage = TRUE)
plot_sex_epiweek(data_grouped, col_names = c("sexo", "semana"), percentage = FALSE)
}
