% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_sex_semanaepi}
\alias{plot_sex_semanaepi}
\title{Generar gráfico de distribución de casos por sexo y semana epidemiológica}
\usage{
plot_sex_semanaepi(
  data_agrupada,
  col_nombres = c("sexo", "semana"),
  fuente_data = NULL
)
}
\arguments{
\item{data_agrupada}{Un data frame que contiene los datos de la enfermedad
o evento agrupados}

\item{col_nombres}{Un array (arreglo) de character (cadena de caracteres) con
los nombres de columna de los datos agrupados de la enfermedad o evento que
contienen el sexo y las semanas epidemiológicas; su valor por defecto es
c("sexo", "semana")}

\item{fuente_data}{Un character (cadena de caracteres) que contiene la
leyenda o fuente de información de los datos; su valor por defecto es NULL}
}
\value{
Un plot o gráfico de distribución de casos por sexo y semana
epidemiológica
}
\description{
Función que genera el gráfico de distribución de casos por sexo
y semana epidemiológica
}
\examples{
data_event <- import_data_event(2020, "DENGUE")
data_event <- limpiar_data_sivigila(data_event, 2020)
data_agrupada <- agrupar_sex_semanaepi(data_event,
                                       col_nombres = c("sexo", "semana"),
                                       porcentaje = TRUE)
plot_sex_semanaepi(data_agrupada = data_agrupada,
                   col_nombres = c("sexo", "semana"))
}
