% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_special_population}
\alias{plot_special_population}
\title{plot_special_population}
\usage{
plot_special_population(disease_data, col_name = "poblacion", percentage = T)
}
\arguments{
\item{disease_data}{Disease data}

\item{col_name}{Data set column name}
}
\value{
A plot of cases distribution by special population
}
\description{
Función que genera la gráficas de distribucion de casos por poblacion especial
Function that generates the plot of cases distribution by special population
}
\examples{
disease_data <-  import_data_disease_by_year(2020, "DENGUE")
plot_special_population(disease_data, col_name = "poblacion", percentage = T)
}
