% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_special_population}
\alias{plot_special_population}
\title{Plot cases distribution by special population}
\usage{
plot_special_population(
  data_grouped,
  col_name = "poblacion",
  percentage = FALSE
)
}
\arguments{
\item{data_grouped}{The disease data grouped}

\item{col_name}{Column names in the disease data grouped that contains 
the ages and sex}

\item{percentage}{Indicates if the data has percentages}
}
\value{
A plot of cases distribution by special population
}
\description{
Function that generates the plot of cases distribution by special population
}
\examples{
disease_data <- import_data_disease_by_year(2020, "DENGUE")
disease_data <- clean_header(disease_data)
data_grouped <- group_by_age_and_sex(disease_data, 
                                     col_names = c("edad", "sexo"), 
                                     percentage = TRUE)
plot_special_population(data_grouped, 
                         col_name = "poblacion", 
                         percentage = FALSE)
}
