% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_variable}
\alias{plot_variable}
\title{Plot by variable(s) or column(s)}
\usage{
plot_variable(
  data,
  var_x,
  var_y,
  var_per = NULL,
  var_fill = NULL,
  wt_per = TRUE,
  label_x,
  label_y,
  scale_name = NULL,
  scale_labels = NULL,
  diagram_title = NULL,
  legend_pos,
  bar_wd = 1,
  text_sz = 3,
  show_val = TRUE,
  caption_label = NULL
)
}
\arguments{
\item{data}{The data to plot}

\item{var_x}{The variable name for axis x}

\item{var_y}{The variable name for axis y}

\item{var_per}{The variable name if the data have percentages}

\item{var_fill}{The variable fill}

\item{wt_per}{Indicates if the data has percentages}

\item{label_x}{The label for axis x}

\item{label_y}{The label for axis y}

\item{scale_name}{The scale name}

\item{scale_labels}{The scale labels}

\item{diagram_title}{The diagram title}

\item{legend_pos}{The legend position}

\item{bar_wd}{The bar width}

\item{text_sz}{The text size}

\item{show_val}{Indicates if the bars should show the values}

\item{caption_label}{Caption or information source of the disease data}
}
\value{
The plot by variable(s) or column(s)
}
\description{
Function that generates a graph by any type of variable or column of 
dataframe
}
\examples{
disease_data <- import_linelist_disease_year(2019, "DENGUE")
disease_data <- clean_header(disease_data)
cases_sex <- group_sex(disease_data, 
                       percentage = TRUE)
plot_variable(cases_sex,
  var_x = col_name,
  var_y = "casos",
  var_fill = col_name,
  var_per = "porcentaje",
  label_x = "Sexo",
  label_y = "Numero de casos",
  scale_name = "Sexo",
  scale_labels = c("Femenino", "Masculino"),
  legend_pos = "right",
  bar_wd = 0.5,
  text_sz = 3,
  show_val = percentage
  )
}
