% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{remove_error_dates}
\alias{remove_error_dates}
\title{Remove Error Dates}
\usage{
remove_error_dates(disease_data, col_init = "INI_SIN", col_cmp = "FEC_HOS")
}
\arguments{
\item{disease_data}{Disease data}

\item{col_init}{Column name of the date}

\item{col_cmp}{Column name of comparison value}
}
\value{
The data without the erronous dates
}
\description{
Función que elimina las fechas que son superiores al valor de comparación
Function that removes greater dates than the comparison value
}
\examples{
disease_data <- import_data_disease_by_year(2019, "DENGUE")
remove_error_dates(disease_data, col_init = "INI_SI", col_cmp = "FEC_HOS")
}
