% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{remove_nin_values}
\alias{remove_nin_values}
\title{Remove NIN (NA, Infinitive, NaN) values}
\usage{
remove_nin_values(disease_data, col_name)
}
\arguments{
\item{disease_data}{The disease data}

\item{col_name}{Column name in the disease data to evaluate}
}
\value{
The clean data without NA, Infinitive or NaN values
}
\description{
Function that removes a rows if the selected column
values includes NA, Infinitive or NaN
}
\examples{
disease_data <- import_linelist_disease_year(2019, "DENGUE")
disease_data <- clean_header(disease_data)
remove_nin_values(disease_data, col_name = "edad")
}
