% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_data.R
\name{remove_val_nin}
\alias{remove_val_nin}
\title{Eliminar valores NIN (NA, Infinito, NaN)}
\usage{
remove_val_nin(data_event, nom_col)
}
\arguments{
\item{data_event}{Un data frame que contiene los datos de una
enfermedad o evento}

\item{nom_col}{Un character (cadena de caracteres) que contiene el
nombre de la columna en los datos de una enfermedad o evento a evaluar}
}
\value{
Los datos limpios sin valores NA, Infinito o NaN
}
\description{
Función que elimina filas si los valores de la columna seleccionada
incluyen NA, Infinito o NaN
}
\examples{
data_event <- import_data_event(2019, "DENGUE")
data_event <- limpiar_encabezado(data_event)
remove_val_nin(data_event = data_event, nom_col = "edad")
}
