% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sivirep-package.R
\docType{package}
\name{sivirep-package}
\alias{sivirep}
\alias{sivirep-package}
\title{sivirep: Data Wrangling and Automated Reports from 'SIVIGILA' source}
\description{
Provides functions for data wrangling and automated reports from SIVIGILA source.
}
\author{
\strong{Maintainer}: Geraldine GC3mez MillC!n \email{geralidine.gomez@javeriana.edu.co} (\href{https://orcid.org/0009-0007-8701-0568}{ORCID}) [contributor]

Authors:
\itemize{
  \item Zulma M. CucunubC! \email{zulma.cucunuba@javeriana.edu.co} (\href{https://orcid.org/0000-0002-8165-3198}{ORCID}) [contributor]
}

Other contributors:
\itemize{
  \item Hugo Gruson \email{hugo.gruson+R@normalesup.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [contributor]
  \item Laura GC3mez Bermeo \email{gomezblaura@javeriana.edu.co} (\href{https://orcid.org/0000-0003-4028-2893}{ORCID}) [contributor]
  \item Miguel GC!mez [contributor]
  \item Jennifer MC)ndez-Romero [contributor]
  \item Johan CalderC3n [contributor]
  \item Lady FlC3rez-Tapiero [contributor]
  \item VerC3nica Tangarife-Arredondo [contributor]
}

}
\keyword{internal}
