% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sivirep-package.R
\docType{package}
\name{sivirep-package}
\alias{sivirep}
\alias{sivirep-package}
\title{sivirep: Data Wrangling and Automated Reports from 'SIVIGILA' source}
\description{
Provides functions for data wrangling and automated reports from SIVIGILA source.
}
\keyword{internal}
